### 2024-01-22 #6259 @michaeldumelle
@kthyng I filtered JOSS reviewers by "R" and "Earth Sciences and Ecology"

One reviewer that I noticed was Kathryn Doering for their "Topic Area" of fisheries ecology. These spatial stream network models have natural applications to fisheries.
