### 2024-02-12 b363074
fixed error in SSN_to_SSN2, which occurred when there were no prediction sites
