+++ man/names.SSN.Rd
@@ -1,19 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/names.SSN.R
-\name{names.SSN}
-\alias{names.SSN}
-\title{names SSN object}
-\usage{
-\method{names}{SSN}(x, ...)
-}
-\arguments{
-\item{x}{An SSN object.}
-
-\item{...}{Other arguments. Not used (needed for generic consistency).}
-}
-\value{
-Print variable names to console
-}
-\description{
-Extract and print names from the SSN object
-}
