+++ NAMESPACE
@@ -94,7 +94,6 @@ S3method(model.frame,ssn_glm)
 S3method(model.frame,ssn_lm)
 S3method(model.matrix,ssn_glm)
 S3method(model.matrix,ssn_lm)
-S3method(names,SSN)
 S3method(plot,Torgegram)
 S3method(plot,ssn_glm)
 S3method(plot,ssn_lm)
@@ -152,6 +151,7 @@ export(ssn_glm)
 export(ssn_import)
 export(ssn_import_predpts)
 export(ssn_lm)
+export(ssn_names)
 export(ssn_put_data)
 export(ssn_rbeta)
 export(ssn_rbinom)
