+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: SSN2
 Title: Spatial Modeling on Stream Networks
-Version: 0.1.1
+Version: 0.1.2
 Authors@R: c(
     person(given = "Michael",
            family = "Dumelle",
