### 2024-02-20 #6259 @michaeldumelle
@mikemahoney218 thank you so much! Here are a few reviewers that have expertise in ecology/statistics/spatial statistics. The first few have expertise in fisheries/aquatic ecology/aquatic monitoring, which is particularly relevant to this submission:

- Kathryn Doering (fisheries ecology)
- GitHub user name fabbiologia (fisheries)
- GitHub user name fawda123 (aquatic ecology)
- GitHub user name hansenjohnson (marine ecology)
- GitHub user name hvoltbb (fisheries)
- GitHub user name molgor (spatial statistics/ecology)
- David Yannick
- Jed Long
- Louis-Marvin Sander
- Fernando de Pol Mayer

Two JOSS reviewers for which I declare a conflict of interest:

- Matt Higham (recent co-author)
- Jeff Hollister (work together closely in the same organization)

Please let me know if there is anything else you need from me at this time!
