### 2024-02-21 #6389 @editorialbot
[REVIEW]: SSN2: The next generation of spatial stream network modeling in R
**Submitting author:** @michaeldumelle (<a href="http://orcid.org/0000-0002-3393-5529">Michael Dumelle</a>)
**Repository:** https://github.com/USEPA/SSN2
**Branch with paper.md** (empty if default branch): develop
**Version:** v0.2.0(JOSS)
