### 2024-02-21 #6389 @mikemahoney218
👋🏼 @michaeldumelle, @fernandomayer, @k-doering-NOAA: this is the review thread for the paper. Just about all of our communications will happen here from now on. :smile:

```@editorialbot generate my checklist```

as the top of a new comment in this thread. For best results, don't include anything else in the comment!

We aim for reviews to be completed within about 2-4 weeks. Please let me know if you require some more time.

Please feel free to ping me (@mikemahoney218) if you have any questions/concerns. Thanks again so much for agreeing to review!
