### 2024-03-01 73879bd
update rankMatrix method to qr
