+++ R/get_data_object_glm.R
@@ -37,9 +37,9 @@ get_data_object_glm <- function(formula, ssn.object, family, additive, anisotrop
   dots$contrasts <- attr(X, "contrasts")
   xlevels <- .getXlevels(terms_val, obdata_model_frame)
   # find p
-  p <- as.numeric(Matrix::rankMatrix(X))
+  p <- as.numeric(Matrix::rankMatrix(X, method = "qr"))
   if (p < NCOL(X)) {
-    stop("Perfect collinearities detected in X. Remove redundant predictors.", call. = FALSE)
+    warning("There are perfect collinearities detected in X (the matrix of explanatory variables). This may make the model fit unreliable or may cause an error while model fitting. Consider removing redundant explanatory variables and refitting the model.", call. = FALSE)
   }
   # find sample size
   n <- NROW(X)
