### 2024-03-04 #6389 @mikemahoney218
Hi all! Just wanted to bump this thread now that we're about two weeks into the review window.

@fernandomayer, @k-doering-NOAA, @fawda123 : note that when you post this comment on this thread:
`@editorialbot generate my checklist`

You'll get a checklist generated containing all the elements we're asking you to look over as part of your review. Please let me know if you've got any questions/comments/concerns regarding the review!

@michaeldumelle , I should have mentioned this earlier, but if you want to take a look at those "MISSING DOIs" in the Editorialbot message above -- assuming those DOIs correspond to your actual citations, please go ahead and add them to your bibtex file (eg `doi = "10.1002/2015wr018349"`). No rush, but they'll need to get fixed before we'd accept the paper.
