### 2024-03-04 cdfc56c
updated predict documentation to clarify that the preds object in the fitted model is used for prediction
