### 2024-03-04 93a6650
added several missing DOIs
