+++ inst/joss/paper.bib
@@ -4,6 +4,7 @@
   year = {2013},
   publisher = {Springer, NY},
   url = {https://asdar-book.org/},
+  doi = {10.1007/978-1-4614-7618-4}
 }
 
 @Manual{bivand2020rgeos,
@@ -34,7 +35,8 @@
   title={Statistics for spatial data (Revised edition)},
   author={Cressie, Noel},
   year={1993},
-  publisher={Wiley: Hoboken, NJ}
+  publisher={Wiley: Hoboken, NJ},
+  doi={10.1002/9781119115151}
 }
 
 @article{detenbeck2016spatial,
@@ -45,7 +47,8 @@
   number={8},
   pages={6018--6040},
   year={2016},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2015wr018349}
 }
 
 @Article{dumelle2023spmodel,
@@ -56,8 +59,7 @@
   volume = {18},
   number = {3},
   pages = {1--32},
-  doi = {10.1371/journal.pone.0282524},
-  url = {https://doi.org/10.1371/journal.pone.0282524},
+  doi = {10.1371/journal.pone.0282524}
 }
 
 
@@ -68,7 +70,8 @@
   volume={100},
   pages={103028},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jtherbio.2021.103028}
 }
 
 @article{fuller2022riparian,
@@ -79,7 +82,8 @@
   number={7},
   pages={e13626},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/rec.13626}
 }
 
 @article{hill2016stream,
@@ -90,7 +94,8 @@
   number={1},
   pages={120--128},
   year={2016},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/1752-1688.12372}
 }
 
 @article{isaak2014applications,
@@ -112,7 +117,8 @@
   number={11},
   pages={9181--9205},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2017WR020969}
 }
 
 @article{isaak2017scalable,
@@ -123,18 +129,20 @@
   number={2},
   pages={147--156},
   year={2017},
-  publisher={NRC Research Press}
+  publisher={NRC Research Press},
+  doi={10.1139/cjfas-2016-0247}
 }
 
 @article{kattwinkel2020preparing,
   title={Preparing GIS data for analysis of stream monitoring data: The R package openSTARS},
   author={Kattwinkel, Mira and Sz{\"o}cs, Eduard and Peterson, Erin and Sch{\"a}fer, Ralf B},
-  journal={Plos one},
+  journal={PLOS ONE},
   volume={15},
   number={9},
   pages={e0239237},
   year={2020},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0239237}
 }
 
 
@@ -145,18 +153,8 @@
   volume={63},
   number={19},
   year={2015},
-  publisher={University of California at Los Angeles}
-}
-
-@article{mainali2019review,
-  title={A review of spatial statistical approaches to modeling water quality},
-  author={Mainali, Janardan and Chang, Heejun and Chun, Yongwan},
-  journal={Progress in Physical Geography: Earth and Environment},
-  volume={43},
-  number={6},
-  pages={801--826},
-  year={2019},
[…]
