### 2024-03-04 #6389 @michaeldumelle
@mikemahoney218 I updated the bibtex file, incorporating the aforementioned DOIs and adding a few more. I pushed the changes to the joss branch, and you can review the commit [here](https://github.com/USEPA/SSN2/commit/93a665019dde7f81638a7d46b733c08e26b7eb7b). Please let me know when you need anything else from me. Thank you!
