### 2024-03-11 USEPA/SSN2#11 @fawda123
[JOSS Review] Add JOSS badge to README 
Minor suggestion, but consider adding the JOSS badge to the package readme.  Currently this indicates in review, but would include the full DOI when published.

https://github.com/openjournals/joss-reviews/issues/6389
