### 2024-03-11 USEPA/SSN2#12 @fawda123
[JOSS Review] Add test coverage
Consider adding test covering to your CI, including the badge showing percentage of the package that's covered.  You've got a bunch of tests in `tests/testthat`, but it would be nice to see how much of the package is covered. This should be pretty straightforward with `usethis::use_coverage()`.

https://github.com/openjournals/joss-reviews/issues/6389
