### 2024-03-11 USEPA/SSN2#14 @fawda123
[JOSS Review] Add `helper-data.R` to `tests/testthat`
There are several objects that are duplicated across test files.  You can define global objects used in each test file by adding them to `helper-data.R`.  For example, this code is repeated in a lot of the files:

```r
copy_lsn_to_temp()
temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
mf04p <- ssn_import(
    temp_path,
    predpts = c("pred1km", "CapeHorn", "Knapp"),
    overwrite = TRUE
  )
```
Adding it to `helper-data.R` ensures it's only created once and will reduce the time needed when the tests are run.  You can also load any packages in this file that are used in more than one test file.

https://github.com/openjournals/joss-reviews/issues/6389
