### 2024-03-11 USEPA/SSN2#15 @fawda123
[JOSS Review] Edits to paper
Overall, the writing is very good.  I've got just a few minor suggestions.

* Line 45: Consider adding a few more sentences about the preprocessing that's required outside of R.  Is this relatively straightforward or is there any online material to guide a user through this process?  I'm asking more from the perspective of ease of use.  If this is a tedious step, it might deter use of your package.  If it's easy, then this should be stated so that new users are more likely to adopt SSN2 into their workflows.
* Line 52: "special list" - is this a literal R list?
* Line 55: This may just be my preference, and I don't think the journal has any stance on this, but consider using third person throughout, e.g., "The package is loaded into R using:".  Also, it would be good to state that the package can be installed using `install.packages()` since its on CRAN.
* Line 58: Maybe don't even introduce this example and begin with `copy_lsn_to_temp()`.  Using the original example with `system.file()` implies the contents of the installed package will be manipulated, which is probably something you don't want to do.
* Line 135: Maybe mention that you've created S3 methods for these?  Adding methods to generics is one of the more powerful features available for R developers, so I like pointing this out.
*  Line 163: Consider adding one more plot here showing stream model predictions, e.g.,
    ```r
    ggplot() +
        geom_sf(data = mf04p$edges) +
        geom_sf(data = aug_pred, aes(color = .fitted), size = 2) +
        scale_color_viridis_c(limits = c(-1.5, 17), option = "H") +
        theme_bw()
    ```
     This would be a nice way to round out the overview.

https://github.com/openjournals/joss-reviews/issues/6389
