### 2024-03-11 #6389 @fawda123
@michaeldumelle I've finished my initial review of the package and paper, nice work! These package updates are critical to maintain relevancy of "legacy" software as R continues to develop.  It's nice to see the effort put into this work.  I've added a few issues (https://github.com/USEPA/SSN2/issues/11, https://github.com/USEPA/SSN2/issues/12, https://github.com/USEPA/SSN2/issues/13, https://github.com/USEPA/SSN2/issues/14, https://github.com/USEPA/SSN2/issues/15) in the main repo for your consideration.  I think the biggest ask is updates to your unit tests.  Let me know if you have any questions!
