### 2024-03-12 #6389 @michaeldumelle
@fawda123 thank you so much for the thorough and helpful review! I really appreciate the time you put into this, and I look forward to incorporating all of your feedback once the remaining reviews come in.

@mikemahoney218 SSN2 currently has three branches: 1) main, which is up to date with CRAN; 2) which is the development version that is ahead of CRAN; and 3) joss, which is up to date with main but has the joss paper in it. When incorporating feedback on the software, can I push changes to the development branch with the understanding that the changes will be merged into main alongside the next CRAN update? And when incorporating feedback on the paper, I can push changes to the joss branch?
