### 2024-03-12 #6389 @mikemahoney218
Life will be easier if you can merge the development branch into the JOSS branch -- in particular, I think it will be easier for reviewers if there's one branch containing all of the most up-to-date revisions to the code and the paper (and it will also be better when we move to accepting the package and need an archive and DOI). Is that possible?
