### 2024-03-12 #6389 @michaeldumelle
@mikemahoney218 definitely! How does this plan sound? Once I get reviews, I will address code feedback in the development branch and paper feedback in the joss branch. Once I am done addressing code and paper feedback, I will merge the development branch into the JOSS branch so the joss branch has the most up-to-date revisions to both code and the paper.
