### 2024-03-13 #6389 @mikemahoney218
So long as you can make it clear to reviewers what's living where -- especially as @fernandomayer and @k-doering-NOAA are still going to do their initial pass. It's normal for these reviews to involve a lot of back and forth and conversation, much much more so than a traditional peer review where the reviewers only look at the document when you're done working on it. We're fully expecting something closer to a co-production workflow, where reviewers are looking at your code while you're incorporating changes from their reviews and the other reviewers!

That's why I think it'd be better if you can have one branch that's up to date, rather than splitting changes out across branches. I'm not sure what advantage separating code and paper changes has, though!
