### 2024-03-13 #6389 @michaeldumelle
@mikemahoney218 I had planned to leave the joss branch frozen with the software version at the time of the article's (hopeful) publication, which is why all the hassle with the separate branches (as the development branch will eventually be merged into main/CRAN). Do you recommend I add all the joss files to the development branch and eventually merge it into main, placing all the joss files in .Rbuildignore? Then I can delete the joss branch?
