### 2024-03-13 #6389 @mikemahoney218
I personally think that will be the easiest for everyone to follow. We'll be capturing the software at time of acceptance as a Zenodo archive from your repository (associated with a new GH release), so you'll need to have everything in one place by then anyway, and will have the release tag as a permanent record on your repository of "the software as accepted by JOSS". So I think having everything in one branch, with the paper in `.Rbuildignore`, makes the most sense.
