### 2024-03-13 #6389 @k-doering-NOAA
@mikemahoney218 I wanted to report one potential COI: Jay Ver Hoef and I are both employed by NOAA Fisheries, although at different offices (I'm at the Office of Science and Technology, while Jay is at the Alaska Fisheries Science Center). To my knowledge, we've never collaborated or even met.

Is it ok to continue my review?
