### 2024-03-13 #6389 @mikemahoney218
I think so! The [JOSS documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) says that we can waive a COI if:

> you and a submitter are both employed by the same very large organization but in different units without any knowledge of each other.

This sounds like a clearcut case of that to me! If you're still able to make an impartial assessment of the work, then I think it makes total sense for you to continue your review.
