### 2024-03-13 #6389 @k-doering-NOAA
@mikemahoney218 thanks - I thought it could be waived as well upon seeing that line in the JOSS documentation, but wanted to report just in case. I am able to make an impartial assessment of the work. I'll continue reviewing, then, thanks for confirming!
