### 2024-03-14 USEPA/SSN2#13 @k-doering-NOAA
One idea for testing for specific output is using [Snapshot testing](https://testthat.r-lib.org/reference/expect_snapshot.html).
