### 2024-03-14 USEPA/SSN2#12 @k-doering-NOAA
I ran `covr::package_coverage()` on a clone of the main branch and coverage is indeed high! It would be worth showing off with a badge on the readme.

```r
covr::package_coverage()
SSN2 Coverage: 85.78%
R/SSN_to_SSN2.R: 0.00%
src/init.c: 0.00%
R/pseudoR2.R: 33.33%
R/create_netgeometry.R: 42.86%
R/utils.R: 44.44%
R/ssn_import_predpts.R: 49.23%
R/AIC.R: 50.00%
R/ssn_get_data.R: 50.00%
R/partition.R: 53.33%
R/laploglik.R: 54.55%
R/ssn_import.R: 58.33%
R/checks.R: 59.32%
R/ssn_put_data.R: 61.29%
R/get_anisotropy_corrected.R: 61.90%
R/get_cholprods.R: 65.22%
R/augment.R: 70.37%
R/get_dist_predbk_object.R: 72.50%
R/random.R: 73.33%
R/deviance.R: 75.00%
R/restruct_ssn_missing.R: 75.00%
R/ssn_update_path.R: 75.00%
R/anova.R: 75.34%
R/local.R: 76.32%
R/get_eigenprods.R: 77.27%
R/optim.R: 78.95%
R/randcov_vector.R: 79.41%
R/get_data_object_glm.R: 79.47%
R/predict_glm.R: 79.80%
R/logLik.R: 80.00%
R/ssn_split_predpts.R: 80.00%
R/createBinaryID.R: 80.56%
R/loocv_glm.R: 80.58%
R/get_cov_matrix.R: 81.48%
R/ssn_write.R: 82.61%
R/predict.R: 82.69%
R/residuals.R: 83.33%
R/get_data_object.R: 84.85%
R/ssn_get_netgeometry.R: 85.71%
R/use_laploglik.R: 85.94%
R/plot.R: 86.07%
R/loocv.R: 86.52%
R/tidy.R: 87.08%
R/get_dist_pred_object.R: 87.08%
R/print.SSN.R: 87.50%
R/ssn_rbinom.R: 87.50%
R/ssn_create_distmat.R: 87.69%
R/ssn_glm.R: 88.46%
R/glances.R: 88.89%
src/test_fc_UNIX.c: 89.47%
R/copy_lsn_to_temp.R: 90.00%
R/cov_initial_search_glm.R: 90.82%
R/ssn_subset.R: 90.85%
R/ssn_get_stream_distmat.R: 90.91%
R/ssn_rinvgauss.R: 90.91%
R/covmatrix.R: 91.07%
R/partition_vector.R: 91.30%
R/ssn_rpois.R: 91.67%
R/ssn_rnorm.R: 92.42%
R/fitted.R: 93.33%
R/ssn_rgamma.R: 93.75%
R/ssn_rnbinom.R: 93.75%
R/coef.R: 94.12%
R/get_Torgegram_initial_object.R: 94.44%
R/cov_initial_search.R: 94.51%
R/get_model_stats_glm.R: 95.86%
R/ssn_rbeta.R: 96.15%
R/get_model_stats.R: 96.27%
R/get_cov_vector.R: 96.97%
R/initial.R: 97.37%
R/initial_NA.R: 97.75%
R/laploglik_products.R: 97.95%
R/Torgegram.R: 98.11%
R/ssn_lm.R: 98.39%
R/use_gloglik.R: 98.39%
R/get_dist_object.R: 98.61%
R/print.R: 99.34%
R/amongSitesDistMat.R: 100.00%
R/confint.R: 100.00%
R/cooks.distance.R: 100.00%
R/cov_betahat_adjust.R: 100.00%
R/cov_estimate.R: 100.00%
R/cov_matrix.R: 100.00%
R/cov_vector.R: 100.00%
R/formula.R: 100.00%
R/get.rid.fc.R: 100.00%
R/getObsRelationshipsDF.R: 100.00%
R/getPredRelationshipsDF.R: 100.00%
R/glance.R: 100.00%
R/gloglik.R: 100.00%
R/gloglik_products.R: 100.00%
R/hatvalues.R: 100.00%
R/influence.R: 100.00%
R/labels.R: 100.00%
R/model.frame.R: 100.00%
R/model.matrix.R: 100.00%
R/names.SSN.R: 100.00%
R/optim2orig.R: 100.00%
R/orig2optim.R: 100.00%
R/params.R: 100.00%
R/ssn_simulate.R: 100.00%
R/summary.R: 100.00%
R/summary.SSN.R: 100.00%
R/transform_anis.R: 100.00%
R/varcomp.R: 100.00%
R/vcov.R: 100.00%
```
