### 2024-03-14 USEPA/SSN2#15 @k-doering-NOAA
For simplicity, I'm also adding suggested edits here. None of these are major and feel free to use your discretion to incorporate or not!

line 55: Consider adding installation instructions in the paper.
line 56: packages -> package
line 62: To streamline the example, consider only showing copying and manipulating. To me, it seems like a bad practice to permanently alter package data locally.
code after 62: consider `file.path(tempdir(), "MiddleFork04.ssn")` instead of using `paste0()`
code after line 72: The additive option is not explained.
line 104: consider mentioning that the `tidy()`, `glance()`, and `augment()` functions are from the broom package.
line 135: I don't think these functions return all of the information that tidy, glance and augment do? If that is the case, perhaps add the words "some of" after "obtain"
line 169: Consider saying "the SSN R package" to distinguish between SSN models and the SSN software
line 184: [VAST](https://github.com/James-Thorson-NOAA/VAST) has been used to implement stream network models ([Charsley et al. 2023](https://doi.org/10.1016/j.fishres.2022.106583)), although it does not seem to be tailor-made to quickly work with SSN models. Consider if this example should be incorporated.
