### 2024-03-14 USEPA/SSN2#16 @k-doering-NOAA
Missing documentation
A few small pieces of the Documentation parts of the [checklist](https://github.com/openjournals/joss-reviews/issues/6389#issuecomment-1995867057) are missing:

- The statement of need on the readme is missing who the target audience is (it is hinted at, but not said explicitly)
- For community guidelines, there is no information on the readme about how to seek support (contributing and reporting issues is covered on the readme/by the CONTRIBUTING.md file)
