### 2024-03-14 USEPA/SSN2#17 @k-doering-NOAA
suggestion: roxygen documentation for non-exported functions
This is only a non-blocking suggestion to help with maintenance in the future.

I noticed some functions that aren't exported are missing roxygen documentation. For example,

https://github.com/USEPA/SSN2/blob/6b634c5ac1ce66195db5cbdec6c9ac3d10bee46a/R/amongSitesDistMat.R#L1-L4

The comments on lines 2-4 could instead be turned into roxygen documentation, which would be more explicit about what the function does and what the variables are. This can benefit future maintainers so it is easier to remember (or learn for the first time) the purpose of the function.
