### 2024-03-14 #6389 @k-doering-NOAA
@michaeldumelle et al., fantastic work! A nicely written article and R package. I just completed the first pass through my checklist.

I did feel there were [two small pieces of documentation missing](https://github.com/USEPA/SSN2/issues/16), so I postponed checking 2 of the checkboxes for now. Once the missing documentation has been added, I can check them off.

I added [edits to the paper](https://github.com/USEPA/SSN2/issues/15#issuecomment-1998566155) and [one suggestion](https://github.com/USEPA/SSN2/issues/17), but these are non-blocking. I also made a few comments on issues posted by @fawda123, in case they can be of some help.
