+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added `ssn_names()` to return column names in the `edges`, `obs`, and `preds` elements of an SSN object.
 * Changed `Matrix::rankMatrix(X, method = "tolNorm2")` to `Matrix::rankMatrix(X, method = "qr")` to enhance stability when determining linear independence in `X`, the design matrix of explanatory variables.
 * Replaced an error message with a warning message when `X` has perfect collinearities (i.e., is not full rank).
+* Minor vignette updates.
 * Minor documentation updates.
 
 ## Bug Fixes
