### 2024-04-08 #6389 @mikemahoney218
I think it would make sense to get started! Especially here where neither completed review is asking for fundamental functionality changes (I think), it's unlikely that you're going to get asked to change anything _back_. Ideally JOSS reviews are a bit more like ongoing discussions than the typical long-term back-and-forth of other journals, so it's fine for you to get started even with one review still outstanding.
