### 2024-04-08 USEPA/SSN2#12 @michaeldumelle
Thank you for the great suggestion @fawda123, and thanks @k-doering-NOAA for locally running `covr::package_coverage()`. I would very much love to add this badge, and I recognize the importance of incorporating CI test coverage. Unfortunately, USEPA GitHub does not currently allow their repositories to use certain cloud-based services, one of which is `codecov.io`. So I cannot incorporate the badge, but maybe there is another way to communicate the coverage of the package as indicated by `covr::package_coverage()`? Also tagging @mikemahoney218. Thank you!
