### 2024-04-08 2320092
Add JOSS badge to README.
