### 2024-04-08 014a99f
add helper data to testthat
