+++ tests/testthat/test-extras-ssn_lm.R
@@ -1,20 +1,6 @@
-test_that("blank test", {
-  expect_null(NULL)
-})
-
 test_local <- FALSE # FALSE for CRAN
 
 if (test_local) {
-  # Copy the mf04p .ssn data to a local directory and read it into R
-  # When modeling with your .ssn object, you will load it using the relevant
-  # path to the .ssn data on your machine
-  copy_lsn_to_temp()
-  temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
-  mf04p <- ssn_import(
-    temp_path,
-    predpts = c("pred1km", "CapeHorn", "Knapp"),
-    overwrite = TRUE
-  )
 
   ssn_create_distmat(
     ssn.object = mf04p,
