+++ tests/testthat/test-utils.R
@@ -1,17 +1,6 @@
-# Copy the mf04p .ssn data to a local directory and read it into R
-# When modeling with your .ssn object, you will load it using the relevant
-# path to the .ssn data on your machine
-copy_lsn_to_temp()
-temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
-mf04p <- ssn_import(
-  temp_path,
-  predpts = c("pred1km", "CapeHorn", "Knapp"),
-  overwrite = TRUE
-)
+# test some utility functions
 
-# fit an example model
-ssn_mod <- ssn_lm(Summer_mn ~ ELEV_DEM, mf04p, tailup_type = "exponential", additive = "afvArea")
-
+# get an initial value object
 initial_object_val <- get_initial_object(
   tailup_type = "exponential",
   taildown_type = "exponential",
@@ -22,6 +11,11 @@ initial_object_val <- get_initial_object(
   euclid_initial = NULL,
   nugget_initial = NULL
 )
+
+# fit an example model
+ssn_mod <- ssn_lm(Summer_mn ~ ELEV_DEM, mf04p, tailup_type = "exponential", additive = "afvArea")
+
+
 
 ################################################################################
 ############################ check_optim_method
