+++ tests/testthat/test-ssn_simulate.R
@@ -1,14 +1,4 @@
 test_that("simulating works", {
-  # Copy the mf04p .ssn data to a local directory and read it into R
-  # When modeling with your .ssn object, you will load it using the relevant
-  # path to the .ssn data on your machine
-  copy_lsn_to_temp()
-  temp_path <- paste0(tempdir(), "/MiddleFork04.ssn")
-  mf04p <- ssn_import(
-    temp_path,
-    predpts = c("pred1km", "CapeHorn", "Knapp"),
-    overwrite = TRUE
-  )
 
   tu <- tailup_params("exponential", de = 1, range = 1)
   td <- taildown_params("exponential", de = 1, range = 1)
