### 2024-04-08 USEPA/SSN2#14 @michaeldumelle
Thanks @fawda123; this is a great suggestion that will be so helpful in my other software packages too! Incorporated in [this commit](https://github.com/USEPA/SSN2/commit/014a99f70e6b02db73f7f8497b69881f18610daa) (on the develop branch).
