### 2024-04-08 USEPA/SSN2#16 @michaeldumelle
Thanks @k-doering-NOAA! I added the following to the statement of need in the journal article and in the README (on the develop branch), which now reads "The `SSN2` **R** package is designed to help users fit SSN models to their stream network data." How is this? I also added a `CONTRIBUTING to SSN2` section to the top of the README, just after `Installation Instructions`. The relevant commit is [linked here](https://github.com/USEPA/SSN2/commit/c7671c24f3a2a760a5f0e27bdb9e791092ba7bc4).
