### 2024-04-09 USEPA/SSN2#16 @k-doering-NOAA
@michaeldumelle the target audience looks good!

I still feel like the "seek support" piece is missing from the readme (although how to contribute is covered well) - what should users do if they need help using SSN2 or discover what they think is a bug? Maybe just one sentence on the readme under the contributing section (or a "getting help" section) could address this?
