### 2024-04-10 USEPA/SSN2#16 @michaeldumelle
@k-doering-NOAA I totally agree; thanks for pointing this out! How do the new `Contributing to SSN2` and `Getting Help` sections in the README (on develop) from [this commit](https://github.com/USEPA/SSN2/commit/15aad49b7a34c9f6bd8773727d293c88495ca433) look?
