### 2024-04-11 USEPA/SSN2#12 @fawda123
@michaeldumelle that's a bummer, but I understand the limitations.  It's possible you could fork the repo to your personal account and setup the codecov CI there? Then you could just put the link to the badge from your account on the USEPA readme? This isn't ideal though as it's not really continuous integration - youd' have to make sure that the USEPA repo is always in sync with your personal.  So maybe not... just a suggestion, I don't think you should do this.  You can close this issue if you like.
