### 2024-04-11 USEPA/SSN2#12 @michaeldumelle
@fawda123 thanks, and I agree it is a total bummer! A few of us are in discussions about adding coverage CI to USEPA repositories, so if this becomes feasible in the future, I will certainly add the codecov CI badge to the SSN2 README.
