### 2024-04-12 USEPA/SSN2#15 @fernandomayer
I'll also add only a comment here to avoid open another issue.

I agree with all the comments/suggestions above and just have one question/comment:

line 39: I understand the use of the Euclidian distance here and the role it plays here. However, I would think in another way to consider physical barriers (e.g. river margins) to the spatial modelling. I believe in some cases this should also be useful. One way to do this would be through the INLA/SPDE approach (e.g. https://arxiv.org/pdf/1608.03787.pdf). Have you ever considered it, or it doesn't make sense?

Note: this is just a comment and this discussion doesn't need to be in the article.
