### 2024-04-12 USEPA/SSN2#20 @fernandomayer
[JOSS Review] Comments
Regarding the review being made for JOSS https://github.com/openjournals/joss-reviews/issues/6389, I'll just leave here two comments:

### README

- Maybe cite non-standard packages needed, and why they're needed (e.g. `spmodel`, `RSQLite`).

### Installation

- Installation on my Linux machine runs fine, since the C code in `src` is compiled using the configurations in `{R_HOME}/etc/Makeconf`. However, if I try to load the package using `devtools::load_all()`, the flags are not identified. I would recommend to create a `src/Makevars` and (possibly) a `src/Makevars.win` to be sure the correct flags are used during compilation. This should improve maintenance and future contributions.

- In the `DESCRIPTION` file, base- and recommended-packages doesn't need to be listed in the 'Imports' field, as they already are bundled wit R itself. (Also, this causes duplications of basic function's names in the search path). Apparently, only sf, generics, tibble, spmodel and RSQLite should be listed in 'Imports'.

In general, it's a very nice work and a very useful package. COngratulations.
