### 2024-04-16 #6389 @michaeldumelle
Thank you very much @fawda123 , @k-doering-NOAA , and @fernandomayer for the thoughtful reviews, and I appreciate your contributions to the software and article. @mikemahoney218 , I am working on these and plan to be done within a few weeks -- I will reach out if any questions come up. If anyone has any additional thoughts/recommendations, please don't hesitate to reach out!
