### 2024-04-22 6268017
helpfile for ssn_get_netgeom
