+++ man/ssn_get_netgeom.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/ssn_get_netgeometry.R
+% Please edit documentation in R/ssn_get_netgeom.R
 \name{ssn_get_netgeom}
 \alias{ssn_get_netgeom}
 \title{Extract netgeom column}
