### 2024-04-22 e42d30f
Modified all import/export functions to 1) ensure that both gpkg and shp could be imported and only gpkg exported. 2) predpts is passed in as a named or unnamed vector instead of a list. 3) works with the output from SSNbler. New function get_sf_obj included to provide the flexibility in (1) and (2). Old versions of functions stored in inst/old.
