+++ R/ssn_import_predpts.R
@@ -5,8 +5,8 @@
 #'   \code{SSN}, \code{ssn_lm}, or \code{ssn_glm}.
 #' @param x An object of class\code{SSN}, \code{ssn_lm}, or
 #'   \code{ssn_glm}.
-#' @param predpts Name of the prediction point shapefile to import in
-#'   character format, without the .shp extension.
+#' @param predpts Name of the prediction point dataset to import in
+#'   character format. See details.
 #' @param format_additive Logical indicating whether the columns containing
 #'   the addtive function values should be formated for
 #'   \code{SSN2}. Default = \code{FALSE}.
@@ -15,46 +15,32 @@
 #'   values. Must be defined if \code{format_additive = TRUE}. Default =
 #'   \code{NULL}.
 #'
-#' @details \command{ssn_import_predpts} imports a shapefile of
-#'   prediction points residing in the .ssn directory into an existing
+#' @details \command{ssn_import_predpts} imports one set of prediction
+#'   points residing in the .ssn directory into an existing
 #'   \code{SSN}, \code{ssn_lm}, or \code{ssn_glm} object. The
-#'   prediction dataset must reside in the ssn.object$path
+#'   prediction dataset must be in shapefile or geopackage format
+#'   (.shp or .gpkg, respectively) and reside in the ssn.object$path
 #'   directory. The path for an \code{SSN} object can be updated using
 #'   \command{ssn_update_path()} prior to importing prediction
-#'   datasets. Note that, the prediction dataset must contain the
+#'   datasets. The argument \code{predpts} accepts the name of the
+#'   prediction point dataset, with or without the file extension. If
+#'   it is passed as a named vector (of length 1), then the name
+#'   provided is used as the name for prediction dataset in the
+#'   \code{SSN} object prediction sites list
+#'   (e.g. \code{names(ssn.obj$preds)}). See
+#'   \code{\link[SSN2]{ssn_import}} for a detailed description of the
+#'   prediction dataset format within the \code{SSN} class
+#'   object.
+#'
+#'   The prediction dataset specified in \code{predpts} must contain the
 #'   spatial, topological and attribute information needed to make
 #'   predictions using an ssn_lm or ssn_glm object.  This information
-#'   can be generated using a number of proprietary and open source
-#'   software tools: \itemize{ \item{The Spatial Tools for the
-#'   Analysis of River Systems (STARS) tools for ArcGIS Desktop
-#'   versions 9.3x-10.8x (Peterson and Ver Hoef 2014). This custom
-#'   ArcGIS toolset is designed to work with existing streams data in
-#'   vector format.}  \item{The openSTARS package (Kattwinkel et
-#'   al. 2020) extends the functionality of the STARS toolset, which
-#'   makes use of R and GRASS GIS. It is open source and designed to
-#'   derive streams in raster format from a digital elevation model
-#'   (DEM).}  \item{The SSNbler package (currently in development as
-#'   of September 2023) is an open source version of the STARS
-#'   toolset, which makes use of the functionality found in the sf
-#'   package to process streams data in vector format.}  }
+#'   is generated using the \code{SSNbler} package, which makes use of
+#'   the functionality found in the \code{sf} and \code{igraph}
+#'   packages to process streams data in vector format.
 #' @return an object of class \code{SSN}, \code{ssn_lm}, or
[…]
