+++ inst/old/ssn_split_predpts_old.R
@@ -9,7 +9,7 @@
 #'
 #' @param ssn An \code{SSN} object.
 #' @param predpts A character string representing the name of the
-#'   prediction dataset. 
+#'   prediction dataset
 #' @param size_predpts numeric value representing the size of the new
 #'   prediction sets. The existing prediction set is split equally to
 #'   produce multiple prediction sets of this size
@@ -28,20 +28,20 @@
 #'   levels should be dropped in the \code{by} column when the new
 #'   prediction dataset(s) are created. Default is \code{FALSE}
 #' @param overwrite logical indicating whether the new prediction
-#'   dataset geopackage should be deleted in the .ssn directory if it
+#'   dataset shapefile should be deleted in the .ssn directory if it
 #'   already exists. Default = \code{FALSE}
 #'
 #' @details Three methods have been provided to split prediction sets:
-#'   size, by, and subset. The
-#'   size method is used to split the existing prediction
-#'   set into multiple equally-sized prediction sets using the \code{size_predpts} argument. Note that the
+#'   \code{size_predpts}, \code{by}, and \code{subset}. The
+#'   \code{size_predpts} method is used to split the existing prediction
+#'   set into multiple equally-sized prediction sets. Note that the
 #'   final prediction set may be smaller in size than the others if
 #'   the total number of predictions is not evenly divisible by
-#'   \code{size_predpts}. The by method is used if the prediction
+#'   \code{size_predpts}. The \code{by} method is used if the prediction
 #'   set is to be split into multiple new prediction sets based on an
 #'   existing column of type factor, integer, character, or
-#'   logical specified using the argument \code{by}. The subset method is used to create one new
-#'   prediction set based on a logical expression defined in \code{subset}.
+#'   logical. The \code{subset} method is used to create one new
+#'   prediction set based on a logical expression.
 #'
 #'   When more than one prediction dataset is created the prediction
 #'   dataset names will be appended with a hyphen and prediction
@@ -61,20 +61,19 @@
 #'   predictions can be made.
 #'
 #' @return returns the \code{SSN} specified in \code{ssn}, with one or more new prediction
-#'   sets. Geopackages of the new prediction sets are written to the
+#'   sets. Shapefiles of the new prediction sets are written to the
 #'   .ssn directory designated in ssn$path.
 #'
 #' @name ssn_split_predpts
-#' @export
 #' @examples
 #' ## Import SSN object
 #' copy_lsn_to_temp() ## Only needed for this example
 #' ssn <- ssn_import(paste0(tempdir(), "/MiddleFork04.ssn"),
-#'   predpts = c("pred1km", "Knapp", "CapeHorn"),
+#'   predpts = c("pred1km.shp", "Knapp", "CapeHorn"),
 #'   overwrite = TRUE
 #' )
 #'
-#' ## Split predictions based on 'size' method
+#' ## Split predictions into size_predpts 200
 #' ssn1 <- ssn_split_predpts(ssn, "CapeHorn",
 #'   size_predpts = 200,
 #'   keep = FALSE, overwrite = TRUE
@@ -82,7 +81,7 @@
 #' names(ssn1$preds)
 #' nrow(ssn1$preds[["CapeHorn-1"]])
 #'
-#' ## Split predictions using 'by' method
+#' ## Split predictions using by method
 #' ssn$preds$pred1km$net.fac <- as.factor(ssn$preds$pred1km$netID)
[…]
