### 2024-04-24 f521376
removed arguments in ssn_import_predpts related to the additive function value
