+++ R/ssn_import.R
@@ -101,9 +101,9 @@
 #' ## Import SSN object with 3 sets of prediction sites
 #' mf04p <- ssn_import(paste0(tempdir(), "/MiddleFork04.ssn"),
 #'   predpts = c(
-#'     "pred1km.shp",
-#'     "CapeHorn.shp",
-#'     "Knapp.shp"
+#'     "pred1km",
+#'     "CapeHorn",
+#'     "Knapp"
 #'   ),
 #'   overwrite = TRUE
 #' )
