+++ R/data.R
@@ -4,23 +4,23 @@
 #' and topological information needed to construct an SSN object using
 #' the \code{SSN2} package.
 #'
-#' The \code{MiddleFork04.ssn} folder contains five shapefiles:
+#' The \code{MiddleFork04.ssn} folder contains five geopackages:
 #'
 #' \itemize{
-#'    \item edges: polyline shapefile representing the stream network
-#'    \item sites: point shapefile representing the observed site locations
-#'    \item pred1km: point shapefile representing prediction site locations at
+#'    \item edges: geopackage with LINESTRING geometry representing the stream network
+#'    \item sites: geopackage with POINT geometry representing the observed site locations
+#'    \item pred1km: geopackage with POINT geometry prediction site locations at
 #'      approximately 1km intervals throughout the stream network
-#'    \item Knapp: point shapefile representing prediction site locations on the Knapp River
-#'    \item CapeHorn: point shapefile representing prediction site locations on the Cape Horn River
+#'    \item Knapp: geopackage with POINT geometry representing prediction site locations on the Knapp River
+#'    \item CapeHorn: geopackage with POINT geometry representing prediction site locations on the Cape Horn River
 #' }
 #'
-#' The \code{MiddleFork04.ssn} includes one text file, \code{netID1.txt}, which contains the
-#' topological information for the stream network in the Middle Fork 2004
+#' The \code{MiddleFork04.ssn} includes two text files, \code{netID1.txt} and \code{netID2.txt}, which contain the
+#' topological information for the two stream networks in the Middle Fork 2004
 #' dataset.
 #'
 #' The distance folder contains four folders that store the hydrologic
-#' distance matrices for each of the point shapefiles (\code{obs}, \code{CapeHorn},
+#' distance matrices for each of the point datasets (\code{obs}, \code{CapeHorn},
 #' \code{Knapp}, and \code{pred1km}). See [ssn_create_distmat()] for a
 #' detailed description of the distance matrix file structure.
 #'
@@ -29,16 +29,16 @@
 #'
 #'  \code{edges}:
 #'   \itemize{
+#'     \item rid: Reach identifier
 #'     \item COMID: Common identifier of an NHD feature or relationship
 #'     \item GNIS_Name: Feature name as found in the Geographic Names Information System
 #'     \item REACHCODE: Unique identifier for a reach. The first 8 digits contain the identfier for the HUC8 and the last 6 digits are a unique within-HUC8 identifier for the reach
 #'     \item FTYPE: three-digit integer used to classify hydrography features in the NHD and define subtypes
 #'     \item FCODE: Numeric code that contains the feature type and its attributes as found in the NHDFCode lookup table
-#'     \item CDRAINAG: Cumulative drainage area (km2) for the lowermost location on the edge
 #'     \item AREAWTMAP: Area weighted mean annual precipitation (mm) at the lowermost location on the edge
 #'     \item SLOPE: Slope of the edge (cm/cm)
-#'     \item h2oAreaKm2: Watershed area (km2) for the lowermost location on the line segment
-#'     \item rid: Reach identifier
+#'     \item rcaAreaKm2: Reach contributing area (km2) for each edge feature. The RCA represents the land area that drains directly to the edge feature
+#'     \item h2oAreaKm2: Watershed area (km2) for the lowermost location on the edge feature
 #'     \item areaPI: Segment proportional influence value, calculated using watershed area (h2oAreaKm2)
 #'     \item afvArea: Additive function value, calculated using areaPI
 #'     \item upDist: Distance from the stream outlet (most downstream location in the the stream network) to the uppermost location on the line segment
@@ -47,9 +47,10 @@
 #' }
[…]
