+++ R/ssn_import.R
@@ -9,10 +9,10 @@
 #'   found within the .ssn folder. See details.
 #' @param overwrite default = \code{FALSE}. If \code{TRUE}, overwrite
 #'   existing binaryID.db files.
-#' 
+#'
 #' @details The \command{ssn_import} function imports spatial data (shapefile or geopackage format)
 #'   from a .ssn folder generated using the
-#'   \code{SSNbler} package function \code{\link[SSNbler]{lsn_to_ssn}}. The .ssn folder contains all of the spatial, topological and
+#'   \code{SSNbler} package function \code{SSNbler::lsn_to_ssn}. The .ssn folder contains all of the spatial, topological and
 #'   attribute data needed to fit a spatial statistical stream network
 #'   model to streams data.  This includes:
 #' \itemize{
@@ -80,7 +80,7 @@
 #'     \item{path: The local filepath for the .ssn directory associated with the \code{SSN}
 #'     object.}
 #'   }
-#' 
+#'
 #' @references
 #' Peterson, E., and Ver Hoef, J.M. (2014) STARS: An
 #'   ArcGIS toolset used to calculate the spatial information needed
@@ -111,15 +111,15 @@
 
 ssn_import <- function(path, include_obs = TRUE, predpts = NULL,
                         overwrite = FALSE) {
-  
+
   if(!dir.exists(path)) stop("Cannot find the .ssn folder.")
- 
+
   # Get wd
   old_wd <- getwd()
   on.exit(setwd(old_wd)) # if the function crashes or finishes, it will restore the initial working directory
-  
+
   local_dir(path)
-  
+
   ## Adjust if relative pathname is supplied in path
   if(substr(path, start = 1, stop = 2) == "./") {
     rel.path <- substr(path, start = 2, stop = nchar(path))
@@ -140,16 +140,16 @@ ssn_import <- function(path, include_obs = TRUE, predpts = NULL,
     p.names <- NULL
 
     for(d in 1:length(predpts)) {
-      
+
       shp.ext<- substr(predpts[d], nchar(predpts[d])-3, nchar(predpts[d])) == ".shp"
       gpkg.ext <- substr(predpts[d], nchar(predpts[d])-4, nchar(predpts[d])) == ".gpkg"
-      
+
       p.names[d] <- ifelse(
         shp.ext == TRUE, substr(predpts[d], 1, nchar(predpts[d])-4),
 
                  ifelse(gpkg.ext == TRUE,
                         substr(predpts[d], 1, nchar(predpts[d])-5),
-                        
+
                  ifelse(shp.ext == FALSE & gpkg.ext == FALSE, predpts[d])))
     }
   }
@@ -159,13 +159,13 @@ ssn_import <- function(path, include_obs = TRUE, predpts = NULL,
   if(!is.null(predpts)) {
     predpts<- basename(predpts)
   }
-  
-   
+
+
   ## ----------------------------------------------------
   ## Import edges
   ## ----------------------------------------------------
   sfedges <- get_sf_obj("edges")
-  
+
   if(exists("sfedges")) {
     ## Check geometry type
     if(sum(st_geometry_type(sfedges, by_geometry = TRUE) ==
@@ -177,7 +177,7 @@ ssn_import <- function(path, include_obs = TRUE, predpts = NULL,
     if(!"geometry" %in% colnames(sfedges)) {
       sf::st_geometry(sfedges) <- "geometry"
     }
-    
+
     ## Add network geometry column to edges
     sfedges[, "netgeom"] <-
       paste0("ENETWORK (", paste(
@@ -185,21 +185,21 @@ ssn_import <- function(path, include_obs = TRUE, predpts = NULL,
                              sfedges$rid,
                              sfedges$upDist),
              ")", sep = "")
-    
+
   } else {
[…]
