+++ R/ssn_create_distmat.R
@@ -114,7 +114,7 @@
 #' copy_lsn_to_temp()
 #' ## Import SSN data
 #' mf04p <- ssn_import(paste0(tempdir(), "/MiddleFork04.ssn"),
-#'   predpts = c("pred1km.shp", "Knapp"),
+#'   predpts = c("pred1km.gpkg", "Knapp"),
 #'   overwrite = TRUE
 #' )
 #'
