### 2024-05-01 92a0988
update ssn names to work with zero observed sites
