+++ R/ssn_import.R
@@ -219,7 +219,8 @@ ssn_import <- function(path, include_obs = TRUE, predpts = NULL,
                                      sfsites$ratio, sfsites$pid, sfsites$locID),
                                    ")", sep = "")
   } else {
-    sfsites <- NA
+    # sfsites <- NA
+    sfsites <- list()
   }
 
   ## ----------------------------------------------------
