### 2024-05-01 2ac84fe
updated mf documentation after removing Knapp
