### 2024-05-02 91eac6e
remove knapp from package mentions
