+++ R/ssn_subset.R
@@ -45,7 +45,7 @@
 #' ## Import SSN object
 #' copy_lsn_to_temp() ## Only needed for this example
 #' mf04p <- ssn_import(paste0(tempdir(), "/MiddleFork04.ssn"),
-#'   predpts = c("pred1km", "Knapp"),
+#'   predpts = "pred1km",
 #'   overwrite = TRUE
 #' )
 #'
@@ -65,11 +65,11 @@
 ssn_subset <- function(ssn, path, subset, clip = FALSE, overwrite = FALSE) {
 
   ## Add .ssn extension if necessary
-  if(substr(path, nchar(path)-3, nchar(path)) != ".ssn") {   
+  if(substr(path, nchar(path)-3, nchar(path)) != ".ssn") {
     print(paste0("path must include .ssn extension. ", path, " updated to ", paste0(path, ".ssn")))
     path <- paste0(path, ".ssn")
   }
-  
+
   file <- path
 
   suppressWarnings({
