+++ tests/testthat/test-predict.R
@@ -9,7 +9,7 @@ test_that("prediction works", {
 
   ssn_create_distmat(
     ssn.object = mf04p,
-    predpts = c("CapeHorn", "Knapp"),
+    predpts = c("CapeHorn"),
     overwrite = TRUE,
     only_predpts = TRUE
   )
@@ -22,7 +22,7 @@ test_that("prediction works", {
 
   expect_vector(predict(ssn_mod1, "CapeHorn"))
   expect_vector(predict(ssn_mod1, "pred1km", block = TRUE))
-  expect_true(inherits(predict(ssn_mod1, "Knapp", interval = "confidence"), "matrix"))
+  expect_true(inherits(predict(ssn_mod1, "CapeHorn", interval = "confidence"), "matrix"))
   expect_true(inherits(predict(ssn_mod1, "pred1km", interval = "prediction", level = 0.9), "matrix"))
   expect_type(predict(ssn_mod1), type = "list")
 
@@ -32,7 +32,7 @@ test_that("prediction works", {
     nugget_type = "nugget", additive = "afvArea"
   )
 
-  expect_vector(predict(ssn_mod2, "Knapp", type = "link"))
+  expect_vector(predict(ssn_mod2, "CapeHorn", type = "link"))
   expect_true(inherits(predict(ssn_mod2, "pred1km", interval = "confidence", level = 0.9), "matrix"))
   expect_true(inherits(predict(ssn_mod2, "CapeHorn", interval = "prediction"), "matrix"))
   expect_type(predict(ssn_mod2), type = "list")
