+++ man/ssn_subset.Rd
@@ -58,7 +58,7 @@ matrices before fitting models to the data in the subset
 ## Import SSN object
 copy_lsn_to_temp() ## Only needed for this example
 mf04p <- ssn_import(paste0(tempdir(), "/MiddleFork04.ssn"),
-  predpts = c("pred1km", "Knapp"),
+  predpts = "pred1km",
   overwrite = TRUE
 )
 
