+++ inst/old/ssn_import_old.R
@@ -1,362 +1,0 @@
-#' Import \code{SSN} object
-#'
-#' @description This function reads spatial data from a .ssn folder
-#'   and creates an \code{SSN} object.
-#' @param path Filepath to the .ssn directory. See details.
-#' @param include_obs default = \code{TRUE}. Logical indicating
-#'   whether observed sites should be included in the SSN object.
-#' @param predpts Vector of shapefile basenames for prediction sites
-#'   found within the .ssn folder.
-#' @param format_additive Logical indicating whether the columns containing
-#'   the addtive function values should be formated for
-#'   \code{SSN2}. Default = \code{FALSE}.
-#' @param names_additive Character vector of column names in observed and
-#'   prediction site datasets containing additive function values.
-#'   Must be defined if \code{format_additive = TRUE}. Default =
-#'   \code{NULL}.
-#' @param overwrite default = \code{FALSE}. If \code{TRUE}, overwrite
-#'   existing binaryID.db files.
-#' @details The \command{importSSN} function imports spatial data from a .ssn
-#'   folder to create an \code{SSN} object. The information contained in the
-#'   .ssn folder can be generated using a number of proprietary and
-#'   open source software tools:
-#'   \itemize{
-#'     \item{The Spatial Tools for the Analysis of River Systems
-#'     (STARS) tools for ArcGIS Desktop versions 9.3x-10.8x (Peterson
-#'     and Ver Hoef 2014). This custom ArcGIS toolset is designed to
-#'     work with existing streams data in vector format.}
-#'     \item{The openSTARS package (Kattwinkel et al. 2020) extends
-#'     the functionality of the STARS toolset, which makes use of R
-#'     and GRASS GIS. It is open source and designed to derive streams
-#'     in raster format from a digital elevation model (DEM).}
-#'     \item{The SSNbler package (currently in development as of
-#'     September 2023) is an open source version of the STARS toolset,
-#'     which makes use of the functionality found in the sf package to
-#'     process streams data in vector format.}
-#'   }
-#'   When spatial data are processed using one of these software
-#'   tools, a .ssn directory is output which contains all of the
-#'   spatial, topological and attribute data needed to fit a spatial
-#'   statistical stream network model to streams data.  This includes:
-#'   \itemize{
-#'     \item{An edges shapefile of lines that represent the stream
-#'     network.}
-#'     \item{A sites shapefile of points where observed data were
-#'     collected on the stream network.}
-#'     \item{Prediction sites shapefile(s) of locations where
-#'     predictions will be made.}
-#'     \item{netID.dat files for each distinct network, which store
-#'     the topological relationships of the line segments in edges.}
-#'   }
-#'   A more detailed description of the .ssn directory and its
-#'   contents is provided in Peterson and Ver Hoef (2014).
-#'
-#'   The \command{ssn_import} imports the edges, observed sites, and
-#'   prediction sites as \code{sf data.frame} objects. A new column named 'netgeom'
-#'   is created to store important data that represents
-#'   topological relationships in a spatial stream network
-#'   model. These data are stored in character format, which is less
-#'   likely to be inadvertantly changed by users. See
-#'   \code{\link[SSN2]{ssn_get_netgeom}} for a more detailed description of
-#'   the format and contents of 'netgeom'.
-#'
[…]
