### 2024-05-03 86c6048
update to printing functions to use NA for obs instead of empty list
