### 2024-05-03 71e8c6b
use div references in vignette instead of fenced blocks
