+++ vignettes/introduction.Rmd
@@ -61,11 +61,11 @@ to account for differences in flow volume or other influential variables [@peter
 result, non-zero covariances are restricted to flow-connected sites in a
 tail-up model. In a tail-down model, the moving average function
 points in the downstream direction. In contrast to the tail-up model,
-tail-down models describe both flow-connected and flow-unconnected
-relationships, although covariances will always be equal or stronger
+tail-down models allow for both flow-connected and flow-unconnected
+autocorrelation, autocorrelation will always be equal or stronger
 for flow-unconnected sites than flow-connected sites separated by
 equal stream distances [@ver2010moving]. In the tail-down model, covariance is
-a function of stream distance and weights are not required.
+a function of stream distance and weights are not required. 
 However, it is also possible and often preferable to build
 spatial stream network models based on a mixture of four components: a
 tail-up component, a tail-down component, a Euclidean component, and a
@@ -81,11 +81,11 @@ of spatial stream network models and their covariance components, see
 
 The `SSN2` **R** package is used to fit and summarize spatial stream network models and make predictions at unobserved locations (Kriging). `SSN2` is an updated version of the `SSN` **R** package [@ver2014ssn].  Why did we create `SSN2` to replace `SSN`? There are two main reasons:
 
-1.  `SSN` depends on the `rgdal` [@bivand2021rgdal], `rgeos` [@bivand2020rgeos], and `maptools` [@bivand2021maptools] **R** packages, which are being retired in October, 2023. Their functionality has been replaced and modernized by the `sf` package [@pebesma2018sf]. `SSN2` depends on `sf` instead of `rgdal`, `rgeos`, and `maptools`, reflecting this broader change regarding handling spatial data in **R**.
+1.  `SSN` depends on the `rgdal` [@bivand2021rgdal], `rgeos` [@bivand2020rgeos], and `maptools` [@bivand2021maptools] **R** packages, which were retired in October, 2023. Their functionality has been replaced and modernized by the `sf` package [@pebesma2018sf]. `SSN2` depends on `sf` instead of `rgdal`, `rgeos`, and `maptools`, reflecting this broader change regarding handling spatial data in **R**.
     
-    * See @nowosad2023 for more information regarding the retirement of `rgdal`, `rgeos`, and `maptools`, available at https://geocompx.org//post/2023/rgdal-retirement .
+    * See @nowosad2023 for more information regarding the retirement of `rgdal`, `rgeos`, and `maptools`, available at [this link](https://geocompx.org//post/2023/rgdal-retirement).
 
-2.  There are features we added to `SSN2` that would have been difficult to implement in `SSN` without a massive restructuring of `SSN`'s foundation, so we created a new package. For example, the `SSN` objects in `SSN2` are S3 objects but the `SSN` objects in `SSN` were S4 objects. Additionally, many functions were rewritten and/or repurposed in `SSN2` to use generic functions (e.g., block prediction in `SSN2` is performed using `predict()` while in `SSN` it was performed using `BlockPredict()`).
+2.  There are features added to `SSN2` that would have been difficult to implement in `SSN` without a massive restructuring of `SSN`'s foundation, so a new package was created. For example, the `SSN` objects in `SSN2` are S3 objects but the `SSN` objects in `SSN` were S4 objects. Additionally, many functions were rewritten and/or repurposed in `SSN2` to use generic functions (e.g., block prediction in `SSN2` is performed using `predict()` while in `SSN` it was performed using `BlockPredict()`). Moreover, `SSN2` leverages many of the tools in the `spmodel` **R** package for spatial statistical modeling [@dumelle2023spmodel].
 
 This vignette provides an overview of basic features in `SSN2`. We load `SSN2` by running
 ```{r, eval = FALSE}
[…]
