+++ inst/references.bib
@@ -87,6 +87,17 @@
   publisher={John Wiley \& Sons}
 }
 
+@article{dumelle2023spmodel,
+  title={spmodel: Spatial statistical modeling and prediction in R},
+  author={Dumelle, Michael and Higham, Matt and Ver Hoef, Jay M},
+  journal={Plos one},
+  volume={18},
+  number={3},
+  pages={e0282524},
+  year={2023},
+  publisher={Public Library of Science San Francisco, CA USA}
+}
+
 @book{faraway2016extending,
   title={Extending the linear model with R: generalized linear, mixed effects and nonparametric regression models},
   author={Faraway, Julian J},
