+++ inst/tutorials/demos/SSN2-introduction.R
@@ -1,0 +1,324 @@
+#:::::::::::::::::::::::::::::::::::::
+# Data Processing ####
+#:::::::::::::::::::::::::::::::::::::
+
+
+# Load packages ####
+library(SSN2)
+library(ggplot2)
+library(sf)
+
+# Copy Data Folder ####
+## copy MiddleFork04 to the temporary directory
+copy_lsn_to_temp()
+## Find the path to the MiddleFork04 in the temporary directory
+path <- paste0(tempdir(), "/MiddleFork04.ssn")
+
+# Import MiddleFork04 SSN ####
+## Import data and prediction points
+mf04p <- ssn_import(
+  path = path,
+  predpts = c("pred1km", "CapeHorn"),
+  overwrite = TRUE
+)
+## print a summary
+summary(mf04p)
+## print names of prediction data
+names(mf04p$preds)
+## visualize stream network, prediction data, observed data
+ggplot() +
+  geom_sf(data = mf04p$edges) +
+  geom_sf(data = mf04p$preds$pred1km, pch = 17, color = "blue") +
+  geom_sf(data = mf04p$obs, color = "brown", size = 2) +
+  theme_bw()
+## create distance matrices needed for spatial statistical modeling
+ssn_create_distmat(
+  ssn.object = mf04p,
+  predpts = c("pred1km", "CapeHorn"),
+  among_predpts = TRUE,
+  overwrite = TRUE
+)
+
+#:::::::::::::::::::::::::::::::::::::
+# Statistical Modeling ####
+#:::::::::::::::::::::::::::::::::::::
+
+# Exploratory Visualizations ####
+## Visualize temperature spatially
+ggplot() +
+  geom_sf(data = mf04p$edges) +
+  geom_sf(data = mf04p$obs, aes(color = Summer_mn), size = 2) +
+  scale_color_viridis_c(limits = c(-1.5, 17), option = "H") +
+  theme_bw()
+
+## Visualize relationship between elevation and temperature
+ggplot(data = mf04p$obs, aes(x = ELEV_DEM, y = Summer_mn)) +
+  geom_point(size = 2) +
+  theme_bw()
+
+## Visualize relationship between precipitation and temperature
+ggplot(data = mf04p$obs, aes(x = AREAWTMAP, y = Summer_mn)) +
+  geom_point(size = 2) +
+  theme_bw()
+
+## Create Torgegram
+tg <- Torgegram(
+  formula = Summer_mn ~ ELEV_DEM + AREAWTMAP,
+  ssn.object = mf04p,
+  type = c("flowcon", "flowuncon", "euclid")
+)
+## Visualize Torgegram
+plot(tg)
+
+# Model Building ####
+## Create a model with all possible covariance components
+ssn_mod <- ssn_lm(
+  formula = Summer_mn ~ ELEV_DEM + AREAWTMAP,
+  ssn.object = mf04p,
+  tailup_type = "exponential",
+  taildown_type = "spherical",
+  euclid_type = "gaussian",
+  additive = "afvArea"
+)
+## Summarize the model
+summary(ssn_mod)
+
+## Inspect the variance components
+varcomp(ssn_mod)
+
+## Glance at the model fit
+glance(ssn_mod)
+
+## Fit a model with just taildown and nugget (notice no additive needed)
+ssn_mod2 <- ssn_lm(
+  formula = Summer_mn ~ ELEV_DEM + AREAWTMAP,
+  ssn.object = mf04p,
+  taildown_type = "spherical"
+)
+
+## Fit a model with just nugget (equivalent to lm)
+ssn_mod3 <- ssn_lm(
[…]
