### 2024-05-08 cb20bad
more demo updates
