### 2024-05-08 e4d0725
some demo updates
