### 2024-05-16 84928db
fix relative path bug with predpts in ssn_import
