### 2024-06-25 USEPA/SSN2#15 @michaeldumelle
Thank you @fawda123, @k-doering-NOAA, and @fernandomayer for the thoughtful comments! I will respond to each in this post, which I will edit as I progress through them. Please reach out if you have any questions/comments/concerns!

# Comments from @fawda123

* Line 45: Consider adding a few more sentences about the preprocessing that's required outside of R. Is this relatively straightforward or is there any online material to guide a user through this process? I'm asking more from the perspective of ease of use. If this is a tedious step, it might deter use of your package. If it's easy, then this should be stated so that new users are more likely to adopt SSN2 into their workflows.
    * We have supplemented this preprocessing discussion and included information about the new `SSNbler` R package, which is an R-based version of STARS and is currently available on GitHub [link here](https://github.com/pet221/SSNbler). We mention that `SSNbler` 1) will soon be on CRAN and 2) contains various resources (for example, see [link here](https://pet221.github.io/SSNbler/), [link here](https://pet221.github.io/SSNbler/articles/introduction.html), and [link here](https://github.com/pet221/SSNbler/tree/main/inst/tutorials)) that guide users through various preprocessing steps.
* Line 52: "special list" - is this a literal R list?
    * Yes, we have removed the word "special".
* Line 55: This may just be my preference, and I don't think the journal has any stance on this, but consider using third person throughout, e.g., "The package is loaded into R using:". Also, it would be good to state that the package can be installed using `install.packages()` since its on CRAN.
    * We have changed to third person throughout and added a mention of `install.packages()`.
* Line 58: Maybe don't even introduce this example and begin with `copy_lsn_to_temp()`. Using the original example with `system.file()` implies the contents of the installed package will be manipulated, which is probably something you don't want to do.
    * We have removed the use of `system.file()` and now begin with `copy_lsn_to_temp()`. We clarified that this is only needed when using the example data by stating "The `copy_lsn_to_temp()` function is only used when working with `MiddleFork04.ssn` and generally, `path` should indicate a permanent directory on your machine that points towards your `.ssn` object.".
* Line 135: Maybe mention that you've created S3 methods for these? Adding methods to generics is one of the more powerful features available for R developers, so I like pointing this out.
    * We now point this out, stating "Additionally, `SSN2` accommodates various S3 methods for commonly-used **R** generic functions that operate on model objects. For example, the generic function `summary()` is used to summarize the fitted model..." and make a few other explicit mentions of generic functions used.
* Line 163: Consider adding one more plot here showing stream model predictions, e.g.,

  ```r
  ggplot() +
      geom_sf(data = mf04p$edges) +
      geom_sf(data = aug_pred, aes(color = .fitted), size = 2) +
      scale_color_viridis_c(limits = c(-1.5, 17), option = "H") +
      theme_bw()
  ```
  This would be a nice way to round out the overview.
    * We have added a version of this figure and agree it is a nice way to round out the overview.

# Comments from @k-doering-NOAA

* line 55: Consider adding installation instructions in the paper.
    * We have added installation instructions.
* line 56: packages -> package
    * Fixed.
* line 62: To streamline the example, consider only showing copying and manipulating. To me, it seems like a bad practice to permanently alter package data locally.
    * We now only show copying and manipulating.
* code after 62: consider `file.path(tempdir(), "MiddleFork04.ssn")` instead of using `paste0()`
    * We have made this change in the paper and introductory vignette.
* code after line 72: The additive option is not explained.
    * We have added the following explanation after the modeling code chunk: "The `additive` argument represents an "additive function value (afv)" variable that captures branching in the stream network and is required when modeling the tailup covariance. Cumulative watershed area is commonly used as the additive function value (here, `afvArea` represents cumulative watershed area), but other variables like flow can be used instead. @ver2010moving provide further details regarding additive function values."
* line 104: consider mentioning that the `tidy()`, `glance()`, and `augment()` functions are from the broom package.
    * We now explicitly mention the `broom` **R** package (we still keep the `broom` citation).
* line 135: I don't think these functions return all of the information that tidy, glance and augment do? If that is the case, perhaps add the words "some of" after "obtain"
    *  We have clarified that the methods for these functions are `SS2`-specific, stating "`SSN2` methods for the `tidy()`, `glance()`, and `augment()` generic functions from the `broom` **R** package [@robinson2021broom] are used to inspect the fitted model and provide diagnostics:"
* line 169: Consider saying "the SSN R package" to distinguish between SSN models and the SSN software
    * We have made this change in the paper and introductory vignette.
* line 184: [VAST](https://github.com/James-Thorson-NOAA/VAST) has been used to implement stream network models ([Charsley et al. 2023](https://doi.org/10.1016/j.fishres.2022.106583)), although it does not seem to be tailor-made to quickly work with SSN models. Consider if this example should be incorporated.
    * We have added a mention of `VAST` for stream network models, and the final few lines of the manuscript read "The `rtop` [@skoien2014rtop], `VAST` [@charsley2023catchment], and `SSN2` **R** packages can be used to describe spatial stream network data in **R**, and `SSN2` provides a familiar and expansive set of tools for model fitting, model diagnostics, and spatial prediction of these data. To learn more about `SSN2`, visit our CRAN webpage at [https://CRAN.R-project.org/package=SSN2](https://CRAN.R-project.org/package=SSN2)."

# Comments from @fernandomayer

* line 39: I understand the use of the Euclidian distance here and the role it plays here. However, I would think in another way to consider physical barriers (e.g. river margins) to the spatial modelling. I believe in some cases this should also be useful. One way to do this would be through the INLA/SPDE approach (e.g. https://arxiv.org/pdf/1608.03787.pdf). Have you ever considered it, or it doesn't make sense?
    *  There are various ways one could accommodate the presence of a physical barrier to the spatial modeling. The first is through the mean (i.e., fixed, covariate, predictor, etc.) structure, whereby one could include covariates that capture whether a site is upstream or downstream of a physical barrier, how far away (upstream or downstream) the site is of the barrier, and potential interactions between these new upstream/downstream covariates and other model covariates (e.g., land cover type, slope, etc.). The second is through the spatial covariance structure, whereby one could add a partition factor that captures whether a site is upstream or downstream of a physical barrier. A partition factor enforces zero correlation among sites with different levels of the factor. For example, if a partition factor for a large dam is in the model, observations upstream of the dam can be spatially correlated with one another, observations downstream of the dam can be spatially correlated with one another, but an observation upstream of the dam is uncorrelated with an observation downstream of the damn. While we think this discussion is too nuanced for a short JOSS article, we have added the following description to our introductory vignette:

    "A partition factor is a variable that allows observations to be uncorrelated when they do not share the same value. For example, one may want to partition the model into two networks (despite their adjacency) because of a significant land mass (or similar obstruction). In this case, it may be relevant to assume observations upstream of the land mass are not correlated with observations downstream of the land mass. Incorporating an effect as a partition factor is different from incorporating the effect as fixed (i.e., on the right-hand side of the `formula` argument). Partition factors affect the covariance portion of the model while fixed effects affect the mean portion.

    Partition factors are modeled using a formula that contains a single variable passed to the `partition_factor` argument:

  ```{r}
  ssn_part <- ssn_lm(
    formula = Summer_mn ~ ELEV_DEM + AREAWTMAP,
    ssn.object = mf04p,
    tailup_type = "exponential",
    taildown_type = "spherical",
    euclid_type = "gaussian",
    additive = "afvArea",
    partition_factor = ~ as.factor(netID)
  )
  ssn_part
  ```

    Here, the partition factor ensures that observations from separate stream networks are completely uncorrelated."

# General Comments

We have made several editorial changes throughout that we believe have enhanced the paper.
