### 2024-06-25 bb7c1b4
joss paper edits part 1
