+++ inst/joss/paper.Rmd
@@ -54,31 +54,33 @@ SSN models use a spatial statistical modeling framework [@cressie1993statistics]
 
 # Package Overview
 
-Before fitting SSN models using `SSN2`, stream network and observation data sets must be pre-processed either by using the STARS toolset for ArcGIS Desktop versions 9.3x - 10.8x [@peterson2014stars] or by using the `openSTARS` **R** package [@kattwinkel2020preparing], which leverages open-source GRASS GIS. Pre-processing using STARS or `openSTARS` ends with the creation of a `.ssn` folder, which is non-proprietary and contains all the spatial, topological, and attribute information needed to fit models to data on a stream network using `SSN2`. Relevant files residing in the `.ssn` folder are read into **R** (using `ssn_import()`) and placed into a special list called an SSN object. The SSN object contains geometry and topological information about the stream reaches and sites, as well as observed data and data for prediction at unsampled sites. 
+Before fitting SSN models using `SSN2`, stream network and observation data sets must be pre-processed either by using the STARS toolset for ArcGIS Desktop versions 9.3x - 10.8x [@peterson2014stars], by using the `openSTARS` **R** package [@kattwinkel2020preparing], which leverages open-source GRASS GIS, or by using the `SSNbler` **R** package [@peterson2024SSNbler], a new, R-based version of STARS that is available on GitHub, will soon be available on CRAN, and contains several useful resources that guide users through these pre-processing steps. Pre-processing using either STARS, `openSTARS`, or `SSNbler` ends with the creation of a `.ssn` folder, which is non-proprietary and has all the spatial, topological, and attribute information needed to fit models to data on a stream network using `SSN2`. Relevant files residing in the `.ssn` folder are read into **R** (using `ssn_import()`) and placed into a list and called an SSN object. The SSN object contains geometry and topological information about the stream reaches and sites, as well as observed data and data for prediction at unsampled sites. 
 
-We first load `SSN2` into our current **R** session:
+`SSN2` is first installed from CRAN:
+```{r, eval = FALSE}
+install.packages("SSN2")
+```
+
+Then, `SSN2` is loaded into our current **R** session:
 ```{r, warning = FALSE}
 library(SSN2)
 ```
 
-The `SSN2` packages comes with an example `.ssn` folder called `MiddleFork04.ssn` that represents water temperatures recorded from a stream network in the Middle Fork of the Salmon River in Idaho, USA during 2004. We may store the file path to this example data:
-```{r}
-path <- system.file("lsndata/MiddleFork04.ssn", package = "SSN2")
-```
+The `SSN2` packages comes with an example `.ssn` folder called `MiddleFork04.ssn` that represents water temperatures recorded from a stream network in the Middle Fork of the Salmon River in Idaho, USA during 2004. 
 
-Several functions in `SSN2` for reading and writing data (which we use shortly) directly manipulate the `.ssn` folder. If it is not desirable to directly manipulate the `MiddleFork04.ssn` data installed alongside `SSN2`, `MiddleFork04.ssn` may be copied it into a temporary directory and the relevant path to this alternative location can be stored:
[…]
