+++ inst/joss/paper.bib
@@ -86,6 +86,15 @@
   doi={10.1111/rec.13626}
 }
 
+@manual{garnier2024viridis,
+  title = {{viridis(Lite)} - Colorblind-Friendly Color Maps for R},
+  author = {{Garnier} and {Simon} and {Ross} and {Noam} and {Rudis} and {Robert} and {Camargo} and Antônio Pedro and {Sciaini} and {Marco} and {Scherer} and {Cédric}},
+  year = {2024},
+  note = {viridis package version 0.6.5},
+  url = {https://sjmgarnier.github.io/viridis/},
+  doi = {10.5281/zenodo.4679423},
+}
+
 @article{hill2016stream,
   title={The Stream-Catchment (StreamCat) Dataset: A database of watershed metrics for the conterminous United States},
   author={Hill, Ryan A and Weber, Marc H and Leibowitz, Scott G and Olsen, Anthony R and Thornbrugh, Darren J},
@@ -255,6 +264,14 @@
   pages={1--17},
   year={2014},
   doi={10.18637/jss.v056.i02}
+}
+
+@manual{peterson2024SSNbler,
+  title = {{SSNbler}: Assemble SSN objects in {R}},
+  author = {Erin Peterson and Michael Dumelle and Alan Pearse and Dan Teleki and Jay M. {Ver Hoef}},
+  year = {2024},
+  note = {{R} package version 0.0.1},
+  url = {https://github.com/pet221/SSNbler}
 }
 
 @article{ruesch2012projected,
