### 2024-06-25 764b90a
joss paper edits part 2
