+++ inst/joss/paper.bib
@@ -31,6 +31,16 @@
   url = {https://CRAN.R-project.org/package=rgdal},
 }
 
+@article{charsley2023catchment,
+  title={Catchment-scale stream network spatio-temporal models, applied to the freshwater stages of a diadromous fish species, longfin eel (Anguilla dieffenbachii)},
+  author={Charsley, Anthony R and Gr{\"u}ss, Arnaud and Thorson, James T and Rudd, Merrill B and Crow, Shannan K and David, Bruno and Williams, Erica K and Hoyle, Simon D},
+  journal={Fisheries Research},
+  volume={259},
+  pages={106583},
+  year={2023},
+  publisher={Elsevier}
+}
+
 @book{cressie1993statistics,
   title={Statistics for spatial data (Revised edition)},
   author={Cressie, Noel},
