### 2024-06-25 3fff328
joss paper edits part 3
