+++ inst/joss/paper.bib
@@ -7,6 +7,17 @@
   doi = {10.1007/978-1-4614-7618-4}
 }
 
+@article{bates2015lme4,
+  title = {Fitting Linear Mixed-Effects Models Using {lme4}},
+  author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and Steve Walker},
+  journal = {Journal of Statistical Software},
+  year = {2015},
+  volume = {67},
+  number = {1},
+  pages = {1--48},
+  doi = {10.18637/jss.v067.i01},
+}
+
 @Manual{bivand2020rgeos,
   title = {{rgeos}: Interface to Geometry Engine - Open Source ('GEOS')},
   author = {Roger Bivand and Colin Rundel},
@@ -284,6 +295,13 @@
   url = {https://github.com/pet221/SSNbler}
 }
 
+@book{pinheiro2006mixed,
+  title={Mixed-effects models in S and S-PLUS},
+  author={Pinheiro, Jos{\'e} and Bates, Douglas},
+  year={2006},
+  publisher={Springer science \& business media}
+}
+
 @article{ruesch2012projected,
   title={Projected climate-induced habitat loss for salmonids in the John Day River network, Oregon, USA},
   author={Ruesch, Aaron S and Torgersen, Christian E and Lawler, Joshua J and Olden, Julian D and Peterson, Erin E and Volk, Carol J and Lawrence, David J},
@@ -405,4 +423,11 @@
   isbn = {978-3-319-24277-4},
   url = {https://ggplot2.tidyverse.org},
   doi={10.1007/978-0-387-98141-3}
+}
+
+@book{zimmerman2024spatial,
+  title={Spatial Linear Models for Environmental Data},
+  author={Zimmerman, Dale L and Ver Hoef, Jay M},
+  year={2024},
+  publisher={CRC Press}
 }
