+++ vignettes/introduction.Rmd
@@ -531,7 +531,9 @@ ssn_rand
 
 `random = ~ as.factor(netID)` is short-hand for `random = ~ (1 | as.factor(netID))`, which is the more familiar `lme4` or `nlme` syntax.
 
-A partition factor is a variable that allows observations to be uncorrelated when they from different levels of the partition factors. For example, one may want to partition the model into two networks despite their adjacency because of a significant land mass or similar obstruction. Partition factors are modeled using a formula that contains a single variable passed to the `partition_factor` argument. We model the two networks as uncorrelated by running
+A partition factor is a variable that allows observations to be uncorrelated when they do not share the same value. For example, one may want to partition the model into two networks (despite their adjacency) because of a significant land mass (or similar obstruction). In this case, it may be relevant to assume observations upstream of the land mass are not correlated with observations downstream of the land mass. Incorporating an effect as a partition factor is different from incorporating the effect as fixed (i.e., on the right-hand side of the `formula` argument). Partition factors affect the covariance portion of the model while fixed effects affect the mean portion. 
+
+Partition factors are modeled using a formula that contains a single variable passed to the `partition_factor` argument:
 ```{r}
 ssn_part <- ssn_lm(
   formula = Summer_mn ~ ELEV_DEM + AREAWTMAP,
@@ -545,7 +547,9 @@ ssn_part <- ssn_lm(
 ssn_part
 ```
 
-In short, the partition factor enables model fitting that builds independence in places not typical on a stream network but deemed relevant by the researcher.
+Here, the partition factor ensures that observations from separate stream networks are completely uncorrelated.
+
+<!-- resulting from Euclidean distance (these observations already do not share any tail-up or tail-down covariance because they are from separate stream networks). -->
 
 ## Generalized Linear SSN Models {#sec:glm}
 
