### 2024-06-26 USEPA/SSN2#20 @michaeldumelle
Thank you for the feedback @fernandomayer ! I will respond to each comment in this post, which I will edit as I progress through them. Please reach out if you have any questions/comments/concerns!

* Maybe cite non-standard packages needed, and why they're needed (e.g. `spmodel`, `RSQLite`).

    * We have added two sections to the `README`: Imported Packages and Suggested Packages. In each section, we briefly detail why the package is imported (or suggested) in `SSN2`.

* Installation on my Linux machine runs fine, since the C code in `src` is compiled using the configurations in `{R_HOME}/etc/Makeconf`. However, if I try to load the package using `devtools::load_all()`, the flags are not identified. I would recommend to create a `src/Makevars` and (possibly) a `src/Makevars.win` to be sure the correct flags are used during compilation. This should improve maintenance and future contributions.

    * We have one small (a few lines) function that uses C. Because 1) Linux users can still download the package, 2) we expect most of our users to request changes via Issues (in which we will carry out the development instead of the user), and 3) our lack of experience creating compiler flags, we have decided to keep default `devtools::load_all()` behavior as it pertains to `Makevars` for now. Our primary concern is to avoid accidentally inserting an error into the standard build process from `install.packages()` or `remotes::install()`. We are open to a user developing a `Makevars` file that we could integrate, as long as the flags were sufficiently explained to us (it seems one possible approach is outlined [here](https://groups.google.com/g/rdevtools/c/eCaooGj3sVg); and some more general advice is [here](https://cran.rstudio.com/doc/manuals/r-devel/R-exts.html#Portable-C-and-C_002b_002b-code), [here](https://stackoverflow.com/questions/13183390/rcpp-how-to-reload-recompiled-c-code-without-restarting-r), and [here](https://bookdown.dongzhuoer.com/hadley/r-pkgs/src)).

* In the `DESCRIPTION` file, base- and recommended-packages doesn't need to be listed in the 'Imports' field, as they already are bundled wit R itself. (Also, this causes duplications of basic function's names in the search path). Apparently, only sf, generics, tibble, spmodel and RSQLite should be listed in 'Imports'.

    * [This section](https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Package-Dependencies) in "Writing R Extensions" mentions that declaring a dependency for `base` is not required but does not specifically address the other base-R package (e.g., `graphics`, `methods`, `stats`, `utils`, etc.). The DESCRIPTION files from various commonly-used R packages ([ggplot2](https://github.com/tidyverse/ggplot2/blob/main/DESCRIPTION), [data.table](https://github.com/Rdatatable/data.table/blob/master/DESCRIPTION), [sf](https://github.com/r-spatial/sf/blob/main/DESCRIPTION), [lme4](https://github.com/lme4/lme4/blob/master/DESCRIPTION)) do declare dependencies for these other base-R packages, so I will err on the side of caution and keep them. If you have a specific reference to documentation that clarifies all this, I would love to see it!
