### 2024-06-26 USEPA/SSN2#17 @michaeldumelle
Thank you for the feedback @k-doering-NOAA! I will respond in this post which I will edit as I progress through documenting non-exported functions. Please reach out if you have any questions/comments/concerns!

We have added roxygen documentation for all non-exported functions, which includes functions in:

* `amongSitesDistMat.R`
* `anova.R`
* `checks.R`
* `cov_betahat_adjust.R`
* `cov_initial_search.R`
* `cov_initial_search_glm.R`
* `cov_matrix.R`
* `cov_vector.R`
* `create_netgeometry.R`
* `createBinaryID.R`
* `get_anisotropy_corrected.R`
* `get_cholprods.R`
* `get_cov_matrix.R`
* `get_cov_vector.R`
* `get_data_object.R`
* `get_data_object_glm.R`
* `get_dist_object.R`
* `get_dist_pred_object.R`
* `get_dist_predbk_object.R`
* `get_eigenprods.R`
* `get_model_stats.R`
* `get_model_stats_glm.R`
* `get_sf_obj.R`
* `get_Torgegram_initial_object.R`
* `get.rid.fc.R`
* `getObsRelationshipsDF.R`
* `getPredRelationshipsDF.R`
* `gloglik_products.R`
* `gloglik.R`
* `initial.R`
* `initial_NA.R`
* `laploglik.R`
* `laploglik_products.R`
* `local.R`
* `loocv.R`
* `loocv_glm.R`
* `optim.R`
* `orig2optim.R`
* `optim2orig.R`
* `params.R`
* `partition.R`
* `predict.R`
* `predict_glm.R`
* `random.R`
* `restruct_ssn_missing.R`
* `Torgegram.R`
* `transform_anis.R`
* `use_gloglik.R`
* `use_laploglik.R`
* `utils.R`
