### 2024-06-26 USEPA/SSN2#13 @michaeldumelle
Thank you for the feedback @fawda123 and @k-doering-NOAA! I will respond in this post which I will edit as I progress through adding testing functionality. Please reach out if you have any questions/comments/concerns!

We have updated our unit testing infrastructure to include many unit tests that verify output as intended. The following files in `tests/testthat` have been updated:

- `helper-data.R`
- `test-dist_objects.R`
- `test-extras.R`
- `test-predict.R`
- `test-ssn_glm.R`
- `test-ssn_lm.R`
- `test-ssn_simulate.R`
- `test-ssn-object.R`
- `test-Torgegram.R`
- `test-utils.R`
