+++ R/SSN2.R
@@ -7,7 +7,6 @@
 #' @importFrom RSQLite dbConnect dbDisconnect dbExistsTable dbListTables dbReadTable dbRemoveTable dbWriteTable SQLite
 #' @importFrom generics tidy glance augment
 #' @importFrom graphics abline legend par points title
-#' @importFrom parallel detectCores makeCluster parLapply stopCluster
 #' @importFrom sf st_as_sf st_bbox st_centroid st_coordinates st_crs st_delete st_drop_geometry
 #'   st_geometry_type st_intersects st_read st_set_crs st_set_geometry st_write
 #' @importFrom spmodel AICc covmatrix dispersion_initial dispersion_params glances loocv pseudoR2 randcov_initial randcov_params varcomp
@@ -20,3 +19,5 @@
 #' @importFrom utils read.table tail
 #' @importFrom withr local_dir
 NULL
+
+# #' @importFrom parallel detectCores makeCluster parLapply stopCluster
