+++ NAMESPACE
@@ -205,10 +205,6 @@ importFrom(graphics,legend)
 importFrom(graphics,par)
 importFrom(graphics,points)
 importFrom(graphics,title)
-importFrom(parallel,detectCores)
-importFrom(parallel,makeCluster)
-importFrom(parallel,parLapply)
-importFrom(parallel,stopCluster)
 importFrom(sf,st_as_sf)
 importFrom(sf,st_bbox)
 importFrom(sf,st_centroid)
