+++ inst/joss/paper.md
@@ -44,7 +44,7 @@ editor_options:
 
 # Summary
 
-The `SSN2` **R** package provides tools for spatial statistical modeling, parameter estimation, and prediction on stream (river) networks. `SSN2` is the successor to the `SSN` **R** package [@ver2014ssn], which was archived alongside broader changes in the **R**-spatial ecosystem [@nowosad2023] that included 1) the retirement of `rgdal` [@bivand2021rgdal], `rgeos` [@bivand2020rgeos], and `maptools` [@bivand2021maptools] and 2) the lack of active development of `sp` [@bivand2013applied]. `SSN2` maintains compatibility with the input data file structures used by `SSN` but leverages modern **R**-spatial tools like `sf` [@pebesma2018sf] and provides many useful features that were not available in `SSN`, including new modeling and helper functions, updated fitting algorithms, and simplified syntax consistent with other **R** generic functions.
+The `SSN2` **R** package provides tools for spatial statistical modeling, parameter estimation, and prediction on stream (river) networks. `SSN2` is the successor to the `SSN` **R** package [@ver2014ssn], which was archived alongside broader changes in the **R**-spatial ecosystem [@nowosad2023] that included 1) the retirement of `rgdal` [@bivand2021rgdal], `rgeos` [@bivand2020rgeos], and `maptools` [@bivand2021maptools] and 2) the lack of active development of `sp` [@bivand2013applied]. `SSN2` maintains compatibility with the input data file structures used by the `SSN` **R** package but leverages modern **R**-spatial tools like `sf` [@pebesma2018sf]. `SSN2` also provides many useful features that were not available in the `SSN` **R** package, including new modeling and helper functions, updated fitting algorithms, and simplified syntax consistent with other **R** generic functions.
 
 # Statement of Need
 
@@ -68,13 +68,13 @@ Then, `SSN2` is loaded into our current **R** session:
 library(SSN2)
 ```
 
-The `SSN2` packages comes with an example `.ssn` folder called `MiddleFork04.ssn` that represents water temperatures recorded from a stream network in the Middle Fork of the Salmon River in Idaho, USA during 2004. 
+The `SSN2` package comes with an example `.ssn` folder called `MiddleFork04.ssn` that represents water temperatures recorded from a stream network in the Middle Fork of the Salmon River in Idaho, USA during 2004. 
 
 Several functions in `SSN2` for reading and writing data (which we use shortly) directly manipulate the `.ssn` folder. As to avoid directly manipulating the `MiddleFork04.ssn` data installed alongside `SSN2`, `MiddleFork04.ssn` is instead be copied it into a temporary directory and the relevant path to directory stored:
 
 ```r
 copy_lsn_to_temp()
-path <- paste0(tempdir(), "/MiddleFork04.ssn")
+path <- file.path(tempdir(), "MiddleFork04.ssn")
 ```
 
 The `copy_lsn_to_temp()` function is only used when working with `MiddleFork04.ssn` and generally, `path` should indicate a permanent directory on your machine that points towards your `.ssn` object. After specifying `path`, the stream reaches, observed sites, and prediction sites (`pred1km`) are imported:
@@ -83,7 +83,7 @@ The `copy_lsn_to_temp()` function is only used when working with `MiddleFork04.s
 mf04p <- ssn_import(path, predpts = "pred1km")
 ```
 
-The stream network, observed sites, and prediction sites (Figure$~$\ref{fig:steam-network})  are visualized [@wickham2016ggplot2]:
+The stream network, observed sites, and prediction sites (Figure$~$\ref{fig:steam-network})  are visualized:
 
 ```r
 library(ggplot2)
@@ -109,7 +109,7 @@ Prior to statistical modeling, the `.ssn` object must be supplemented with hydro
[…]
