### 2024-06-28 33478f9
roxygen for non-exported functions
