### 2024-07-02 #6389 @fawda123
@michaeldumelle I'm sure the community is divided on this, but you might consider leveraging chatGPT or other generative AI to help with the unit tests.  It's saved me a lot of time and I've found it an extremely useful tool for creating tests that actually evaluate functionality.  Just copy/paste a function with the Roxygen comments and ask it to write unit tests.
