+++ tests/testthat/test-ssn_simulate.R
@@ -8,8 +8,8 @@ test_that("simulating works", {
   rand <- spmodel::randcov_params("netID" = 1)
 
   # mean seq
-  set.seed(0)
-  mean_seq <- rnorm(n = NROW(mf04p$obs), 0, sd = 0.25)
+  n_obs <- NROW(mf04p$obs)
+  mean_seq <- rnorm(n = n_obs, 0, sd = 0.25)
 
   # set netID as factor
   mf04p$obs$netID <- as.factor(mf04p$obs$netID)
@@ -18,26 +18,43 @@ test_that("simulating works", {
   pf <- ~netID
 
   # ssn_rnorm
-  expect_vector(ssn_simulate(
+  set.seed(0)
+  sim1 <- ssn_simulate(
     family = "gaussian", ssn.object = mf04p, network = "obs",
     tu, td, eu, nug, additive = afvArea, mean = 0, samples = 1
-  ))
-  expect_true(inherits(ssn_simulate(
+  )
+
+  expect_equal(length(sim1), n_obs)
+  expect_equal(sim1[1:2], c(-2.131, -3.127), tolerance = 0.01)
+
+  sim2 <- ssn_simulate(
     family = Gaussian, ssn.object = mf04p, network = "obs",
     tu, td, eu2, nug, additive = "afvArea", mean = mean_seq, samples = 2,
     randcov_params = rand, partition_factor = pf
-  ), "matrix"))
+  )
+
+  expect_equal(dim(sim2), c(n_obs, 2))
+  expect_equal(sim2[1, ], c(3.221, -1.821), tolerance = 0.01)
+
 
   # ssn_rpois
-  expect_vector(ssn_simulate(
+  set.seed(0)
+  sim1 <- ssn_simulate(
     family = "poisson", ssn.object = mf04p, network = "obs",
     tu, td, eu, nug, additive = afvArea, mean = 0, samples = 1
-  ))
-  expect_true(inherits(ssn_simulate(
+  )
+
+  expect_equal(length(sim1), n_obs)
+  expect_equal(sim1[1:2], c(19, 1), tolerance = 0.01)
+
+  sim2 <- ssn_simulate(
     family = poisson, ssn.object = mf04p, network = "obs",
     tu, td, eu2, nug, additive = "afvArea", mean = mean_seq, samples = 2,
     randcov_params = rand, partition_factor = pf
-  ), "matrix"))
+  )
+
+  expect_equal(dim(sim2), c(n_obs, 2))
+  expect_equal(sim2[1, ], c("1" = 3, "2" = 6), tolerance = 0.01)
 
   # ssn_rnbinom
   expect_vector(ssn_simulate(

[…]
