+++ tests/testthat/test-ssn_simulate.R
@@ -8,6 +8,7 @@ test_that("simulating works", {
   rand <- spmodel::randcov_params("netID" = 1)
 
   # mean seq
+  set.seed(0)
   n_obs <- NROW(mf04p$obs)
   mean_seq <- rnorm(n = n_obs, 0, sd = 0.25)
 
@@ -25,7 +26,7 @@ test_that("simulating works", {
   )
 
   expect_equal(length(sim1), n_obs)
-  expect_equal(sim1[1:2], c(-2.131, -3.127), tolerance = 0.01)
+  expect_equal(sim1[1:2], c(2.525, -0.652), tolerance = 0.01)
 
   sim2 <- ssn_simulate(
     family = Gaussian, ssn.object = mf04p, network = "obs",
@@ -34,7 +35,7 @@ test_that("simulating works", {
   )
 
   expect_equal(dim(sim2), c(n_obs, 2))
-  expect_equal(sim2[1, ], c(3.221, -1.821), tolerance = 0.01)
+  expect_equal(sim2[1, ], c(-2.066, 3.221), tolerance = 0.01)
 
 
   # ssn_rpois
@@ -57,61 +58,79 @@ test_that("simulating works", {
   expect_equal(sim2[1, ], c("1" = 3, "2" = 6), tolerance = 0.01)
 
   # ssn_rnbinom
-  expect_vector(ssn_simulate(
-    family = "nbinomial", ssn.object = mf04p, network = "obs",
-    tu, td, eu, nug, additive = afvArea, mean = 0, samples = 1,
-    dispersion = 1
-  ))
-  expect_true(inherits(ssn_simulate(
-    family = nbinomial, ssn.object = mf04p, network = "obs",
-    tu, td, eu2, nug, additive = "afvArea", mean = mean_seq, samples = 2,
-    dispersion = 1, randcov_params = rand, partition_factor = pf
-  ), "matrix"))
-
-  # ssn_rbinom
-  expect_vector(ssn_simulate(
+  set.seed(0)
+  sim1 <- ssn_simulate(
     family = "binomial", ssn.object = mf04p, network = "obs",
     tu, td, eu, nug, additive = afvArea, mean = 0, samples = 1
-  ))
-  expect_true(inherits(ssn_simulate(
+  )
+
+  expect_equal(length(sim1), n_obs)
+  expect_equal(sim1[1:2], c(1, 0), tolerance = 0.01)
+
+  sim2 <- ssn_simulate(
     family = binomial, ssn.object = mf04p, network = "obs",
     tu, td, eu2, nug, additive = "afvArea", mean = mean_seq, samples = 2,
     randcov_params = rand, partition_factor = pf
-  ), "matrix"))
+  )
+
+  expect_equal(dim(sim2), c(n_obs, 2))
+  expect_equal(sim2[1, ], c("1" = 1, "2" = 1), tolerance = 0.01)
 
   # ssn_rbeta
-  expect_vector(ssn_simulate(
+  set.seed(0)
+  sim1 <- ssn_simulate(
     family = "beta", ssn.object = mf04p, network = "obs",
-    tu, td, eu, nug, additive = afvArea, mean = 0, samples = 1,
-    dispersion = 1
-  ))
-  expect_true(inherits(ssn_simulate(
+    tu, td, eu, nug, additive = afvArea, mean = 0, samples = 1
+  )
+
+  expect_equal(length(sim1), n_obs)
+  expect_equal(sim1[1:2], c(0.9997, 0.7529), tolerance = 0.01)
+
+  sim2 <- ssn_simulate(
     family = beta, ssn.object = mf04p, network = "obs",
     tu, td, eu2, nug, additive = "afvArea", mean = mean_seq, samples = 2,
[…]
