### 2024-07-08 df77ae5
some more testing updates
