+++ tests/testthat/test-ssn-object.R
@@ -1,41 +1,99 @@
 test_that("ssn manipulation functions work", {
 
-  expect_no_error(copy_lsn_to_temp())
 
+  # test class of original SSN object
   expect_s3_class(mf04p, "SSN")
 
+
+  tempdir_path <- paste0(tempdir(), "/MiddleFork04.ssn")
+
+  # copy lsn to temporary directory (does it exist?)
+  copy_lsn_to_temp()
+  expect_true(file.exists(tempdir_path))
+  # are the correct files present?
+  file_names <- c("binaryID.db", "CapeHorn.gpkg", "edges.gpkg", "netID1.dat",
+                  "netID2.dat", "pred1km.gpkg", "sites.gpkg")
+  expect_true(all(file_names %in% list.files(tempdir_path)))
+
+
+  # import new SSN object from temporary LSN
   mf04p_new <- ssn_import(paste0(tempdir(), "/MiddleFork04.ssn"),
                       predpts = c("pred1km"),
                       overwrite = TRUE
 
   )
-  expect_no_error(ssn_import_predpts(mf04p_new, predpts = "CapeHorn"))
+  expect_s3_class(mf04p_new, "SSN")
+  expect_identical(names(mf04p_new), c("edges", "obs", "preds", "path"))
+  expect_s3_class(mf04p_new$edges, "sf")
+  expect_s3_class(mf04p_new$obs, "sf")
+  expect_s3_class(mf04p_new$preds$pred1km, "sf")
+  expect_identical(mf04p_new$path, tempdir_path)
 
-  expect_no_error(ssn_update_path(mf04p, mf04p$path))
+  # import capehorn prediction data
+  mf04p_new <- ssn_import_predpts(mf04p_new, predpts = "CapeHorn")
+  expect_s3_class(mf04p_new$preds$CapeHorn, "sf")
 
-  expect_message(ssn_create_distmat(mf04p, predpts = c("pred1km")))
-  expect_no_error(ssn_create_distmat(mf04p, overwrite = TRUE))
-  expect_no_error(ssn_create_distmat(mf04p, predpts = c("pred1km"),
-                     overwrite = TRUE, among_predpts = TRUE, only_predpts = TRUE))
+  # update path
+  mf04p <- ssn_update_path(mf04p, tempdir_path)
+  expect_identical(mf04p$path, tempdir_path)
+
+  # create distance matrices
+  ssn_create_distmat(mf04p_new, predpts = c("pred1km"), overwrite = TRUE)
+  distance_path <- paste0(tempdir_path, "\\distance")
+  expect_true(file.exists(distance_path))
+  expect_true(all(c("obs", "pred1km") %in% list.files(distance_path)))
+  ssn_create_distmat(mf04p_new, predpts = c("pred1km"),
+                     overwrite = TRUE, among_predpts = TRUE, only_predpts = TRUE)
+  obs_path <- paste0(distance_path, "\\obs")
+  expect_true(file.exists(obs_path))
+  obs_files <- c("dist.net1.RData", "dist.net2.RData")
+  expect_true(all(obs_files %in% list.files(obs_path)))
+  pred1km_path <- paste0(distance_path, "\\pred1km")
+  expect_true(file.exists(pred1km_path))
+  pred1km_files <- c("dist.net1.RData", "dist.net2.RData", "dist.net1.a.RData",
+                     "dist.net1.b.RData", "dist.net2.a.RData", "dist.net2.b.RData")
+  expect_true(all(pred1km_files %in% list.files(pred1km_path)))
 
 
+  # subset SSN data
+  subset_path <- paste0(tempdir(), "/subset1.ssn")
[…]
