### 2024-07-08 cb39a97
some test updates
