+++ tests/testthat/test-ssn_lm.R
@@ -1,11 +1,12 @@
 test_that("generics work ssn_lm point data", {
 
-  ssn_mod1 <- ssn_lm(Summer_mn ~ ELEV_DEM, mf04p,
+  form <- Summer_mn ~ ELEV_DEM
+  ssn_mod1 <- ssn_lm(form, mf04p,
     tailup_type = "exponential",
     taildown_type = "exponential", euclid_type = "exponential",
     nugget_type = "nugget", additive = "afvArea"
   )
-  ssn_mod2 <- ssn_lm(Summer_mn ~ ELEV_DEM, mf04p,
+  ssn_mod2 <- ssn_lm(form, mf04p,
     tailup_type = "exponential",
     taildown_type = "none", euclid_type = "none",
     nugget_type = "nugget", additive = "afvArea"
@@ -13,69 +14,109 @@ test_that("generics work ssn_lm point data", {
 
   # AIC
   expect_vector(AIC(ssn_mod1))
-  expect_s3_class(AIC(ssn_mod1, ssn_mod2), "data.frame") # turn reml fixed effects warning off
+  expect_equal(AIC(ssn_mod1), 84.934, tolerance = 0.01)
+  expect_s3_class(AIC(ssn_mod1, ssn_mod2), "data.frame")
+  expect_equal(AIC(ssn_mod1, ssn_mod2)$AIC, c(84.934, 82.893), tolerance = 0.01)
 
   # anova
-  expect_s3_class(anova(ssn_mod1), "data.frame")
-  expect_s3_class(anova(ssn_mod1), "anova.ssn_lm")
-  expect_s3_class(tidy(anova(ssn_mod1)), "data.frame")
-  expect_s3_class(anova(ssn_mod1, ssn_mod2), "data.frame")
-  expect_s3_class(anova(ssn_mod1, ssn_mod2), "anova.ssn_lm")
-  expect_s3_class(tidy(anova(ssn_mod1, ssn_mod2)), "data.frame")
+  anova1 <- anova(ssn_mod1)
+  expect_s3_class(anova1, "data.frame")
+  expect_s3_class(anova1, "anova.ssn_lm")
+  expect_s3_class(tidy(anova1), "data.frame")
+  expect_equal(tidy(anova1)$statistic, c(30.6, 21.1), tolerance = 0.01)
+  anova12 <- anova(ssn_mod1, ssn_mod2)
+  expect_s3_class(anova12, "data.frame")
+  expect_s3_class(anova12, "anova.ssn_lm")
+  expect_s3_class(tidy(anova12), "data.frame")
+  expect_equal(tidy(anova12)$statistic, 5.96, tolerance = 0.01)
 
   # augment
-  expect_s3_class(augment(ssn_mod1), "data.frame")
-  expect_s3_class(augment(ssn_mod1, newdata = "pred1km"), "data.frame")
+  aug_ssn_mod1 <- augment(ssn_mod1)
+  expect_s3_class(aug_ssn_mod1, "sf")
+  expect_equal(aug_ssn_mod1$.fitted[1], 14.261, tolerance = 0.01)
+  aug_pred_ssn_mod1 <- augment(ssn_mod1, newdata = "pred1km")
+  expect_s3_class(aug_pred_ssn_mod1, "sf")
+  expect_equal(aug_pred_ssn_mod1$.fitted[1], 14.690, tolerance = 0.01)
 
   # coef
   expect_vector(coef(ssn_mod1))
+  expect_equal(coef(ssn_mod1), c("(Intercept)" = 70.54, "ELEV_DEM" = -0.0289), tolerance = 0.01)
   expect_s3_class(coef(ssn_mod1, type = "tailup"), "tailup_exponential")
[…]
