+++ tests/testthat/test-ssn_lm.R
@@ -1,5 +1,7 @@
 test_that("generics work ssn_lm point data", {
 
+  set.seed(2)
+
   form <- Summer_mn ~ ELEV_DEM
   ssn_mod1 <- ssn_lm(form, mf04p,
     tailup_type = "exponential",
@@ -33,9 +35,11 @@ test_that("generics work ssn_lm point data", {
   # augment
   aug_ssn_mod1 <- augment(ssn_mod1)
   expect_s3_class(aug_ssn_mod1, "sf")
+  expect_true(all(c(".fitted", ".resid", ".hat", ".cooksd", ".std.resid") %in% names(aug_ssn_mod1)))
   expect_equal(aug_ssn_mod1$.fitted[1], 14.261, tolerance = 0.01)
   aug_pred_ssn_mod1 <- augment(ssn_mod1, newdata = "pred1km")
   expect_s3_class(aug_pred_ssn_mod1, "sf")
+  expect_true(all(c(".fitted") %in% names(aug_ssn_mod1)))
   expect_equal(aug_pred_ssn_mod1$.fitted[1], 14.690, tolerance = 0.01)
 
   # coef
@@ -86,6 +90,8 @@ test_that("generics work ssn_lm point data", {
   expect_equal(fitted(ssn_mod1, type = "tailup")[1], c("1" = -0.130), tolerance = 0.01)
   expect_vector(fitted(ssn_mod1, type = "taildown"))
   expect_equal(fitted(ssn_mod1, type = "taildown")[1], c("1" = 0.772), tolerance = 0.01)
+  expect_vector(fitted(ssn_mod1, type = "euclid"))
+  expect_equal(fitted(ssn_mod1, type = "euclid")[1], c("1" = 0.003), tolerance = 0.01)
   expect_null(fitted(ssn_mod1, type = "randcov"))
   # fitted.values alias
   expect_vector(fitted.values(ssn_mod1))
@@ -94,6 +100,8 @@ test_that("generics work ssn_lm point data", {
   expect_equal(fitted.values(ssn_mod1, type = "tailup")[1], c("1" = -0.130), tolerance = 0.01)
   expect_vector(fitted.values(ssn_mod1, type = "taildown"))
   expect_equal(fitted.values(ssn_mod1, type = "taildown")[1], c("1" = 0.772), tolerance = 0.01)
+  expect_vector(fitted.values(ssn_mod1, type = "euclid"))
+  expect_equal(fitted.values(ssn_mod1, type = "euclid")[1], c("1" = 0.003), tolerance = 0.01)
   expect_null(fitted.values(ssn_mod1, type = "randcov"))
 
   # formula
@@ -116,35 +124,46 @@ test_that("generics work ssn_lm point data", {
   glance_ssn_mod12 <- glances(ssn_mod1, ssn_mod2)
   expect_s3_class(glances(ssn_mod1, ssn_mod2), "data.frame")
   expect_equal(dim(glance_ssn_mod12), c(2, 10))
-  expect_identical(names_glance, names(glance_ssn_mod12))
+  expect_identical(c("model", names_glance), names(glance_ssn_mod12))
 
   # hatvalues
   expect_vector(hatvalues(ssn_mod1))
+  expect_equal(hatvalues(ssn_mod1)[1], c("1" = 0.068), tolerance = 0.01)
 
   # influence
-  expect_s3_class(influence(ssn_mod1), "data.frame")
+  infl_ssn_mod1 <- influence(ssn_mod1)
+  names_infl <- c(".resid", ".hat", ".cooksd", ".std.resid")
+  expect_s3_class(infl_ssn_mod1, "data.frame")
+  expect_equal(dim(infl_ssn_mod1), c(45, 4))
[…]
