+++ tests/testthat/test-ssn_glm.R
@@ -1,11 +1,14 @@
 test_that("generics work ssn_glm point data", {
 
-  ssn_mod1 <- ssn_glm(Summer_mn ~ ELEV_DEM,
+  set.seed(2)
+
+  form <- Summer_mn ~ ELEV_DEM
+  ssn_mod1 <- ssn_glm(form,
     family = "Gamma", mf04p, tailup_type = "exponential",
     taildown_type = "exponential", euclid_type = "exponential",
     nugget_type = "nugget", additive = "afvArea"
   )
-  ssn_mod2 <- ssn_glm(Summer_mn ~ ELEV_DEM,
+  ssn_mod2 <- ssn_glm(form,
     family = Gamma, mf04p, tailup_type = "exponential",
     taildown_type = "none", euclid_type = "none",
     nugget_type = "nugget", additive = "afvArea"
@@ -13,99 +16,160 @@ test_that("generics work ssn_glm point data", {
 
   # AIC
   expect_vector(AIC(ssn_mod1))
-  expect_s3_class(AIC(ssn_mod1, ssn_mod2), "data.frame") # turn reml fixed effects warning off
+  expect_equal(AIC(ssn_mod1), 181.112, tolerance = 0.01)
+  expect_s3_class(AIC(ssn_mod1, ssn_mod2), "data.frame")
+  expect_equal(AIC(ssn_mod1, ssn_mod2)$AIC, c(181.112, 181.888), tolerance = 0.01)
 
   # anova
-  expect_s3_class(anova(ssn_mod1), "data.frame")
-  expect_s3_class(anova(ssn_mod1), "anova.ssn_glm")
-  expect_s3_class(tidy(anova(ssn_mod1)), "data.frame")
-  expect_s3_class(anova(ssn_mod1, ssn_mod2), "data.frame")
-  expect_s3_class(anova(ssn_mod1, ssn_mod2), "anova.ssn_glm")
-  expect_s3_class(tidy(anova(ssn_mod1, ssn_mod2)), "data.frame")
+  anova1 <- anova(ssn_mod1)
+  expect_s3_class(anova1, "data.frame")
+  expect_s3_class(anova1, "anova.ssn_glm")
+  expect_s3_class(tidy(anova1), "data.frame")
+  expect_equal(tidy(anova1)$statistic, c(44.9, 19.9), tolerance = 0.01)
+  anova12 <- anova(ssn_mod1, ssn_mod2)
+  expect_s3_class(anova12, "data.frame")
+  expect_s3_class(anova12, "anova.ssn_glm")
+  expect_s3_class(tidy(anova12), "data.frame")
+  expect_equal(tidy(anova12)$statistic, 8.78, tolerance = 0.01)
 
   # augment
-  expect_s3_class(augment(ssn_mod1), "data.frame")
-  expect_s3_class(augment(ssn_mod1, newdata = "pred1km"), "data.frame")
+  aug_ssn_mod1 <- augment(ssn_mod1)
+  expect_s3_class(aug_ssn_mod1, "sf")
+  expect_true(all(c(".fitted", ".resid", ".hat", ".cooksd", ".std.resid") %in% names(aug_ssn_mod1)))
+  expect_equal(aug_ssn_mod1$.fitted[1], 14.918, tolerance = 0.01)
+  aug_pred_ssn_mod1 <- augment(ssn_mod1, newdata = "pred1km")
+  expect_s3_class(aug_pred_ssn_mod1, "sf")
+  expect_true(all(c(".fitted") %in% names(aug_ssn_mod1)))
+  expect_equal(aug_pred_ssn_mod1$.fitted[1], 2.687, tolerance = 0.01)
 
   # coef
[…]
