+++ inst/joss/paper.Rmd
@@ -23,7 +23,7 @@ authors:
 affiliations:
  - name: Office of Research and Development, United States Environmental Protection Agency (USEPA)
    index: 1
- - name: EP Consulting and Queensland University of Technology, Centre for Data Science Brisbane Australia 4000 
+ - name: EP Consulting and Centre for Data Science, Queensland University of Technology, Brisbane Australia 4000 
    index: 2
  - name: NMFS Alaska Fisheries Science Center, United States National Oceanic and Atmospheric Administration (NOAA)
    index: 3
@@ -32,8 +32,8 @@ affiliations:
  - name: Rocky Mountain Research Station, United States Forest Service (USFS)
    index: 5
 citation_author: Dumelle et. al.
-date: 11 November 2023
-year: 2023
+date: 07 July 2024
+year: 2024
 bibliography: paper.bib
 output: rticles::joss_article
 csl: apa.csl
@@ -44,43 +44,42 @@ editor_options:
 
 # Summary
 
-The `SSN2` **R** package provides tools for spatial statistical modeling, parameter estimation, and prediction on stream (river) networks. `SSN2` is the successor to the `SSN` **R** package [@ver2014ssn], which was archived alongside broader changes in the **R**-spatial ecosystem [@nowosad2023] that included 1) the retirement of `rgdal` [@bivand2021rgdal], `rgeos` [@bivand2020rgeos], and `maptools` [@bivand2021maptools] and 2) the lack of active development of `sp` [@bivand2013applied]. `SSN2` maintains compatibility with the input data file structures used by the `SSN` **R** package but leverages modern **R**-spatial tools like `sf` [@pebesma2018sf]. `SSN2` also provides many useful features that were not available in the `SSN` **R** package, including new modeling and helper functions, updated fitting algorithms, and simplified syntax consistent with other **R** generic functions.
+The `SSN2` **R** package provides tools for spatial statistical modeling, parameter estimation, and prediction on stream (river) networks. `SSN2` is the successor to the `SSN` **R** package [@ver2014ssn], which was archived alongside broader changes in the **R**-spatial ecosystem [@nowosad2023] that included 1) the retirement of `rgdal` [@bivand2021rgdal], `rgeos` [@bivand2020rgeos], and `maptools` [@bivand2021maptools] and 2) the lack of active development of `sp` [@bivand2013applied]. `SSN2` maintains compatibility with the input data file structures used by the `SSN` **R** package but leverages modern **R**-spatial tools like `sf` [@pebesma2018sf]. `SSN2` also provides many useful features that were not available in the `SSN` **R** package, including new modeling and helper functions, enhanced fitting algorithms, and simplified syntax consistent with other **R** generic functions.
 
 # Statement of Need
 
 Streams provide vital aquatic services that sustain wildlife, provide drinking and irrigation water, and support recreational and cultural activities.  Data are often collected at various locations on a stream network and used to characterize spatial patterns in stream phenomena. For example, a manager may need to know how the amount of a hazardous chemical changes throughout a stream network to inform mitigation efforts. Comprehensive formulations of spatial stream network (SSN) models are provided by @ver2010moving, @peterson2010mixed, and @ver2014ssn. The `SSN2` **R** package is designed to help users fit SSN models to their stream network data.
 
[…]
