+++ inst/joss/paper.bib
@@ -1,5 +1,5 @@
 @Book{bivand2013applied,
-  author = {Roger S. Bivand and Edzer Pebesma and Virgilio Gomez-Rubio},
+  author = {Roger Bivand and Edzer Pebesma and Virgilio Gomez-Rubio},
   title = {Applied spatial data analysis with {R}, Second edition},
   year = {2013},
   publisher = {Springer, NY},
@@ -43,7 +43,7 @@
 }
 
 @article{charsley2023catchment,
-  title={Catchment-scale stream network spatio-temporal models, applied to the freshwater stages of a diadromous fish species, longfin eel (Anguilla dieffenbachii)},
+  title={Catchment-scale stream network spatio-temporal models, applied to the freshwater stages of a diadromous fish species, longfin eel (\textit{Anguilla dieffenbachii})},
   author={Charsley, Anthony R and Gr{\"u}ss, Arnaud and Thorson, James T and Rudd, Merrill B and Crow, Shannan K and David, Bruno and Williams, Erica K and Hoyle, Simon D},
   journal={Fisheries Research},
   volume={259},
@@ -108,7 +108,7 @@
 }
 
 @manual{garnier2024viridis,
-  title = {{viridis(Lite)} - Colorblind-Friendly Color Maps for R},
+  title = {{viridis(Lite)} - Colorblind-Friendly Color Maps for {R}},
   author = {{Garnier} and {Simon} and {Ross} and {Noam} and {Rudis} and {Robert} and {Camargo} and Antônio Pedro and {Sciaini} and {Marco} and {Scherer} and {Cédric}},
   year = {2024},
   note = {viridis package version 0.6.5},
@@ -130,7 +130,7 @@
 
 @article{isaak2014applications,
   title={Applications of spatial statistical network models to stream data},
-  author={Isaak, Daniel J and Peterson, Erin E and Ver Hoef, Jay M and Wenger, Seth J and Falke, Jeffrey A and Torgersen, Christian E and Sowder, Colin and Steel, E Ashley and Fortin, Marie-Josee and Jordan, Chris E and others},
+  author={Isaak, Daniel J and Peterson, Erin E and {Ver Hoef}, Jay M and Wenger, Seth J and Falke, Jeffrey A and Torgersen, Christian E and Sowder, Colin and Steel, E Ashley and Fortin, Marie-Josee and Jordan, Chris E and others},
   journal={Wiley Interdisciplinary Reviews: Water},
   volume={1},
   number={3},
@@ -141,7 +141,7 @@
 
 @article{isaak2017norwest,
   title={The NorWeST summer stream temperature model and scenarios for the western US: A crowd-sourced database and new geospatial tools foster a user community and predict broad climate warming of rivers and streams},
-  author={Isaak, Daniel J and Wenger, Seth J and Peterson, Erin E and Ver Hoef, Jay M and Nagel, David E and Luce, Charles H and Hostetler, Steven W and Dunham, Jason B and Roper, Brett B and Wollrab, Sherry P and others},
+  author={Isaak, Daniel J and Wenger, Seth J and Peterson, Erin E and {Ver Hoef}, Jay M and Nagel, David E and Luce, Charles H and Hostetler, Steven W and Dunham, Jason B and Roper, Brett B and Wollrab, Sherry P and others},
   journal={Water Resources Research},
   volume={53},
   number={11},
@@ -153,7 +153,7 @@
 
 @article{isaak2017scalable,
   title={Scalable population estimates using spatial-stream-network (SSN) models, fish density surveys, and national geospatial database frameworks for streams},
[…]
