### 2024-07-09 f2ebf4e
error testing updates
