+++ tests/testthat/test-ssn_glm.R
@@ -56,6 +56,8 @@ test_that("generics work ssn_glm point data", {
   expect_vector(coef(ssn_mod1, type = "dispersion"))
   expect_s3_class(coef(ssn_mod1, type = "dispersion"), "Gamma")
   expect_equal(unclass(coef(ssn_mod1, type = "dispersion")), c("dispersion" = 82810.24), tolerance = 0.01)
+  expect_error(coef(ssn_mod1, type = "error"), 'Invalid type argument. The type argument must be "fixed", "ssn", "tailup",  "taildown",  "euclid",  "nugget", "dispersion", or "randcov".')
+
 
   # coefficients alias
   expect_vector(coefficients(ssn_mod1))
@@ -71,6 +73,7 @@ test_that("generics work ssn_glm point data", {
   expect_vector(coefficients(ssn_mod1, type = "dispersion"))
   expect_s3_class(coefficients(ssn_mod1, type = "dispersion"), "Gamma")
   expect_equal(unclass(coefficients(ssn_mod1, type = "dispersion")), c("dispersion" = 82810.24), tolerance = 0.01)
+  expect_error(coefficients(ssn_mod1, type = "error"), 'Invalid type argument. The type argument must be "fixed", "ssn", "tailup",  "taildown",  "euclid",  "nugget", "dispersion", or "randcov".')
 
   # confint
   expect_equal(dim(confint(ssn_mod1)), c(2, 2))
@@ -85,6 +88,7 @@ test_that("generics work ssn_glm point data", {
   expect_equal(dim(covmatrix(ssn_mod1, "pred1km")), c(175, 45))
   expect_equal(dim(covmatrix(ssn_mod1, "pred1km", cov_type = "obs.pred")), c(45, 175))
   expect_equal(dim(covmatrix(ssn_mod1, "pred1km", cov_type = "pred.pred")), c(175, 175))
+  expect_error(covmatrix(ssn_mod1, "pred1km", cov_type = "error"), 'Invalid "cov_type" argument.')
 
   # deviance
   expect_vector(deviance(ssn_mod1))
@@ -100,6 +104,8 @@ test_that("generics work ssn_glm point data", {
   expect_vector(fitted(ssn_mod1, type = "euclid"))
   expect_equal(fitted(ssn_mod1, type = "euclid")[1], c("1" = 3.574e-09), tolerance = 0.01)
   expect_null(fitted(ssn_mod1, type = "randcov"))
+  expect_error(fitted(ssn_mod1, type = "error"), 'Invalid type argument. The type argument must be "response", "tailup",  "taildown",  "euclid",  "nugget", or "randcov".')
+
   # fitted.values alias
   expect_vector(fitted.values(ssn_mod1))
   expect_equal(fitted.values(ssn_mod1)[1], c("1" = 14.918), tolerance = 0.01)
@@ -110,6 +116,8 @@ test_that("generics work ssn_glm point data", {
   expect_vector(fitted.values(ssn_mod1, type = "euclid"))
   expect_equal(fitted.values(ssn_mod1, type = "euclid")[1], c("1" = 3.574e-09), tolerance = 0.01)
   expect_null(fitted.values(ssn_mod1, type = "randcov"))
+  expect_error(fitted.values(ssn_mod1, type = "error"), 'Invalid type argument. The type argument must be "response", "tailup",  "taildown",  "euclid",  "nugget", or "randcov".')
+
 
   # formula
   expect_type(formula(ssn_mod1), "language")
@@ -196,6 +204,7 @@ test_that("generics work ssn_glm point data", {
                c("fit" = 2.635, "lwr" = 2.444, "upr" = 2.827), tolerance = 0.01)
[…]
