### 2024-07-09 2cdd7ec
some additional joss updates
