+++ inst/joss/paper.bib
@@ -360,12 +360,13 @@
   url = {https://CRAN.R-project.org/package=broom},
 }
 
-@Article{sainsbury2002frk,
-  title = {Modelling big, heterogeneous, non-Gaussian spatial and spatio-temporal data using FRK},
-  author = {Sainsbury-Dale, Matthew and Zammit-Mangion, Andrew and Cressie, Noel},
-  journal = {arXiv:2110.02507},
-  year = {2022},
-  doi = {10.48550/arxiv.2110.02507}
+@article{sainsbury2024modeling,
+  title={Modeling Big, Heterogeneous, Non-Gaussian Spatial and Spatio-Temporal Data Using FRK},
+  author={Sainsbury-Dale, Matthew and Zammit-Mangion, Andrew and Cressie, Noel},
+  journal={Journal of Statistical Software},
+  volume={108},
+  pages={1--39},
+  year={2024}
 }
 
 @article{scown2017improving,
