### 2024-07-09 #6389 @michaeldumelle
@mikemahoney218 I have completed my resubmission! A few things to note:

- The `develop` branch contains the development version of `SSN2`. In the `develop` branch is the joss folder, which contains the manuscript. The `develop` branch will be merged into the `main` branch once the development version (0.2.0) is accepted by CRAN. I will submit to CRAN once the JOSS process is wrapped up and, if accepted, update the citation information.
- I have updated the `NEWS.md` to indicate the JOSS updates, and I can continue to edit that document as we move forward through this process. The relevant `NEWS.md` entry for the JOSS updates reads: "Significant testing, documentation, and auxiliary (e.g., `README.md`) updates as part of a submission to *Journal of Statistical Software*. Relevant issues associated with the review are available at [#11](https://github.com/USEPA/SSN2/issues/11), [#12](https://github.com/USEPA/SSN2/issues/12), [#13](https://github.com/USEPA/SSN2/issues/13), [#14](https://github.com/USEPA/SSN2/issues/14), [#15](https://github.com/USEPA/SSN2/issues/15), [#16](https://github.com/USEPA/SSN2/issues/16), [#17](https://github.com/USEPA/SSN2/issues/17), [#20](https://github.com/USEPA/SSN2/issues/20). The review is [linked here](https://github.com/openjournals/joss-reviews/issues/6389)."

Please let me know if you have any questions/comments/concerns regarding any of my review responses or resubmission!
