### 2024-07-15 USEPA/SSN2#21 @fawda123
final minor text edits
I noticed two typos on lines 59 ('creatino') and 60 ('palced') in the revised JOSS draft.  There also are two instances where first-person is still used on lines 68 ('which we use shortly') and 79 ('wich we will model').  https://github.com/openjournals/joss-reviews/issues/6389
