### 2024-07-15 #6389 @fawda123
@michaeldumelle great job addressing my comments and those from @k-doering-NOAA and @fernandomayer.  I'm very pleased with the revised draft and updates on the dev branch.  In particular, the code testing looks a lot better and I sincerely appreciate the time you invested making these more comprehensive.  It's tedious work but definitely provides peace of mind.  The updated draft is also looking good and I especially like the final figure that was added.  There are a few very minor text edits that I've requested addressed in https://github.com/USEPA/SSN2/issues/21, but otherwise it's looking great!  It's good to go once these final edits are made!
