### 2024-07-15 c7186aa
joss minor text edits
