+++ inst/joss/paper.md
@@ -54,7 +54,7 @@ SSN models use a spatial statistical modeling framework [e.g., @cressie1993stati
 
 # Package Overview
 
-The streams, observation, and prediction datasets must be pre-processed prior to fitting SSN models and making predictions at unobserved locations using `SSN2`. Previously, the STARS toolset for ArcGIS Desktop versions 9.3x - 10.8x [@peterson2014stars] or the `openSTARS` **R** package [@kattwinkel2020preparing] were used to generate spatial information required for model fitting and prediction. However, both software packages have recently been retired and are replaced by the `SSNbler` **R** package [@peterson2024SSNbler], which is a new, **R**-based version of the STARS tools. `SSNbler` is currently available on GitHub, will soon be available on CRAN, and contains several useful resources that guide users through these pre-processing steps. Pre-processing using either `SSNbler`, STARS, or `openSTARS` ends with the creatino of a `.ssn` folder, which is non-proprietary. Files residing in the `.ssn` folder are read into R using `ssn_import()` from `SSN2` and palced into a list structure called an SSN object, which contains all the spatial, topological, and attribute information needed to leverage the modeling tools in `SSN2`.
+The streams, observation, and prediction datasets must be pre-processed prior to fitting SSN models and making predictions at unobserved locations using `SSN2`. Previously, the STARS toolset for ArcGIS Desktop versions 9.3x - 10.8x [@peterson2014stars] or the `openSTARS` **R** package [@kattwinkel2020preparing] were used to generate spatial information required for model fitting and prediction. However, both software packages have recently been retired and are replaced by the `SSNbler` **R** package [@peterson2024SSNbler], which is a new, **R**-based version of the STARS tools. `SSNbler` is currently available on GitHub, will soon be available on CRAN, and contains several useful resources that guide users through these pre-processing steps. Pre-processing using either `SSNbler`, STARS, or `openSTARS` ends with the creation of a `.ssn` folder, which is non-proprietary. Files residing in the `.ssn` folder are read into R using `ssn_import()` from `SSN2` and placed into a list structure called an SSN object, which contains all the spatial, topological, and attribute information needed to leverage the modeling tools in `SSN2`.
 
 `SSN2` is first installed from CRAN:
 
@@ -70,7 +70,7 @@ library(SSN2)
 
 The `SSN2` package comes with an example `.ssn` folder called `MiddleFork04.ssn` that represents water temperatures recorded from a stream network in the Middle Fork of the Salmon River in Idaho, USA during 2004. 
 
-Several functions in `SSN2` for reading and writing data (which we use shortly) directly manipulate the `.ssn` folder. To avoid directly manipulating the `MiddleFork04.ssn` data installed alongside `SSN2`, `MiddleFork04.ssn` is instead copied into a temporary directory and the relevant path to this directory stored:
+Several functions in `SSN2` for reading and writing data directly manipulate the `.ssn` folder. To avoid directly manipulating the `MiddleFork04.ssn` data installed alongside `SSN2`, `MiddleFork04.ssn` is instead copied into a temporary directory and the relevant path to this directory stored:
 
 ``` r
 copy_lsn_to_temp()
@@ -102,14 +102,14 @@ ggplot() +
 \caption{Middle Fork 2004 stream networks. Observed sites are represented by brown, closed circles. Prediction sites are represented by blue, closed triangles.}\label{fig:steam-network}
 \end{figure}
 
-Prior to statistical modeling, hydrologic distance matrices are created: [@ver2010moving]:
+Prior to statistical modeling, hydrologic distance matrices are created [@ver2010moving]:
 
 ``` r
 ssn_create_distmat(mf04p, predpts = "pred1km", overwrite = TRUE)
 ```
 
[…]
