### 2024-07-15 USEPA/SSN2#21 @michaeldumelle
Addressed [in this commit](https://github.com/USEPA/SSN2/commit/c7186aa27579565cf3e14b2df359a053ba48b76f). Thanks for catching these @fawda123 !
