### 2024-07-15 #6389 @michaeldumelle
Thanks for the kind words @fawda123, and I agree the testing infrastructure is now much, much better! I really appreciate the text edits that you suggested in [USEPA/SSN2#21](https://github.com/USEPA/SSN2/issues/21), and they are now incorporated in the draft.

@mikemahoney218 : I am ready to proceed on my end!
