### 2024-07-15 #6389 @mikemahoney218
Awesome! Thank you so much @fawda123 , @fernandomayer , and @k-doering-NOAA for your thoughtful reviews here.

@michaeldumelle you'll see in the post above a checklist with steps for you to work through ("Additional Author Tasks"). At this point could you please work through those steps -- making sure to post the version number of your software release and your archive DOI in this thread, and that the archive has the same metadata (authors, license, and title) as your JOSS publication.

I'll then go ahead with my review of the paper (mostly copy-editing) and then can move forward with recommending acceptance of the submission.
