### 2024-07-15 #6389 @michaeldumelle
Thanks @mikemahoney218 ! A question about the release tag -- which of these three approaches do you recommend?:

1. Generating a release from the current development branch
2. Merging develop into main and generating a release from the main branch
3. Merging develop into main, submitting to CRAN, and generating a release from the main branch
