### 2024-07-15 #6389 @mikemahoney218
Either (1) or (2) seem fine. The primary purpose of the release is so that we can all point to a version and say "this is the one JOSS accepted"; it doesn't need to correspond to anything users would notice. Looking at [recent papers](https://joss.theoj.org/papers/10.21105/joss.06028), I don't think we even publish the version number anywhere on our website.
