### 2024-07-15 #6389 @michaeldumelle
Thanks @mikemahoney218 . How is [this release tag](https://github.com/USEPA/SSN2/releases/tag/v0.2.0(JOSS))? If you approve, I'll use the same one for Zenodo/figshare/etc.
