### 2024-07-15 #6389 @michaeldumelle
@mikemahoney218 I also wanted to check whether it was okay to explicitly thank reviewers/editor in the acknowledgements?
