### 2024-07-15 #6389 @mikemahoney218
@michaeldumelle yes, that's okay! No obligation, but certainly no prohibition either 😄
