### 2024-07-15 #6389 @michaeldumelle
@mikemahoney218 I will wait until I hear back from collaborators about double-checking affiliations and then archive the JOSS materials. If I don't hear from everyone by tomorrow, does that get in the way of your workflow?
