### 2024-07-15 #6389 @mikemahoney218
I won't be able to check your Zenodo until you've got the archive up (of course), which will be the last thing I need to handle before I hand it off to the EiC -- but what I meant to say is that I'll do my checks of the paper for copyediting issues, DOIs and so on in references, and so on, which is usually the thing that involves the most back-and-forth.
