### 2024-07-17 #6389 @mikemahoney218
A few paper comments, all about citation formatting:

- CRAN packages have DOIs now; can you please add the DOIs to packages that don't currently have them
- Bivand et al 2013 -- should "second edition" be removed from the title?
- Charsley 2023 -- Needs `{}` around the genus to preserve capitalization
- Hill 2016 -- Needs `{}` around "United States"
- Kuhn 2022 -- needs `{R}`
- Nychka 2023 -- I think this would be better if it used `{fields}` to preserve the stylization.
- Wickham 2019 -- please capitalize the journal name 😄

Otherwise, once you get the archive out I'll be able to recommend acceptance
