### 2024-07-17 #6389 @michaeldumelle
Thanks @mikemahoney218 I'll get these and a few other affiliation/ORCID updates from coauthors done this week (hopefully today). Then I will push, update the package files in the release tag, archive on Zenodo, and finally give you a notice!
