### 2024-07-17 #6389 @michaeldumelle
@mikemahoney218 can you help me with DOIs from archived packages (e.g., `rgeos`)? Do they follow the same structure as `10.32614/CRAN.package.package_name`? Should I include/omit them in this case?
