### 2024-07-17 #6389 @mikemahoney218
The archived packages don't have DOIs, so they're good to go -- but packages like `fields` do have a DOI on CRAN, even if their default citation file doesn't provide it, and I think it'd be best to include those.
