### 2024-07-17 3adca73
editorial updates to joss
