+++ inst/joss/paper.bib
@@ -1,6 +1,6 @@
 @Book{bivand2013applied,
-  author = {Roger Bivand and Edzer Pebesma and Virgilio Gomez-Rubio},
-  title = {Applied spatial data analysis with {R}, Second edition},
+  author = {Roger Bivand and Edzer Pebesma and Virgilio G{\'o}mez-Rubio},
+  title = {Applied spatial data analysis with {R}},
   year = {2013},
   publisher = {Springer, NY},
   url = {https://asdar-book.org/},
@@ -23,7 +23,7 @@
   author = {Roger Bivand and Colin Rundel},
   year = {2020},
   note = {R package version 0.5-5},
-  url = {https://CRAN.R-project.org/package=rgeos},
+  url = {https://CRAN.R-project.org/package=rgeos}
 }
 
 @Manual{bivand2021maptools,
@@ -31,7 +31,7 @@
   author = {Roger Bivand and Nicholas Lewin-Koh},
   year = {2021},
   note = {R package version 1.1-2},
-  url = {https://CRAN.R-project.org/package=maptools},
+  url = {https://CRAN.R-project.org/package=maptools}
 }
 
 @Manual{bivand2021rgdal,
@@ -39,17 +39,18 @@
   author = {Roger Bivand and Tim Keitt and Barry Rowlingson},
   year = {2021},
   note = {R package version 1.5-27},
-  url = {https://CRAN.R-project.org/package=rgdal},
+  url = {https://CRAN.R-project.org/package=rgdal}
 }
 
 @article{charsley2023catchment,
-  title={Catchment-scale stream network spatio-temporal models, applied to the freshwater stages of a diadromous fish species, longfin eel (\textit{Anguilla dieffenbachii})},
+  title={Catchment-scale stream network spatio-temporal models, applied to the freshwater stages of a diadromous fish species, longfin eel (\textit{{A}nguilla dieffenbachii})},
   author={Charsley, Anthony R and Gr{\"u}ss, Arnaud and Thorson, James T and Rudd, Merrill B and Crow, Shannan K and David, Bruno and Williams, Erica K and Hoyle, Simon D},
   journal={Fisheries Research},
   volume={259},
   pages={106583},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.fishres.2022.106583}
 }
 
 @book{cressie1993statistics,
@@ -117,7 +118,7 @@
 }
 
 @article{hill2016stream,
-  title={The Stream-Catchment (StreamCat) Dataset: A database of watershed metrics for the conterminous United States},
+  title={The Stream-Catchment (StreamCat) Dataset: A database of watershed metrics for the conterminous {United States}},
   author={Hill, Ryan A and Weber, Marc H and Leibowitz, Scott G and Olsen, Anthony R and Thornbrugh, Darren J},
   journal={JAWRA Journal of the American Water Resources Association},
   volume={52},
@@ -136,7 +137,8 @@
   number={3},
   pages={277--294},
   year={2014},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wat2.1023}
 }
 
 @article{isaak2017norwest,
@@ -176,10 +178,10 @@
 }
 
 @book{kuhn2022tidy,
-  title={Tidy modeling with R},
+  title={Tidy modeling with {R}},
   author={Kuhn, Max and Silge, Julia},
   year={2022},
-  publisher={" O'Reilly Media, Inc."}
+  publisher={O'Reilly Media, Inc.}
 }
 
 
@@ -217,7 +219,8 @@
   title={National Stream Internet protocol and user guide},
[…]
