+++ inst/joss/paper.md
@@ -4,7 +4,7 @@ tags:
   - R
   - Spatial stream network
   - Spatial correlation
-  - Spatial prediction
+  - Spatial prediction (Kriging)
   - Geostatistics
   - Generalized linear model
   - Torgegram
@@ -13,23 +13,26 @@ authors:
     orcid: 0000-0002-3393-5529
     affiliation: "1" # (Multiple affiliations must be quoted)
   - name: Erin E. Peterson
+    orcid: 0000-0003-2992-0372
     affiliation: 2
   - name: Jay M. Ver Hoef
+    orcid: 0000-0003-4302-6895
     affiliation: 3
   - name: Alan Pearse
+    orcid: 0000-0002-4133-8548 
     affiliation: 4
   - name: Daniel J. Isaak
     affiliation: 5
 affiliations:
- - name: Office of Research and Development, United States Environmental Protection Agency (USEPA)
+ - name: Pacific Ecological Systems Division, United States Environmental Protection Agency, Corvallis, OR, USA
    index: 1
- - name: EP Consulting and Centre for Data Science, Queensland University of Technology, Brisbane Australia 4000 
+ - name: EP Consulting and Centre for Data Science, Queensland University of Technology, Brisbane, QLD, Australia
    index: 2
- - name: NMFS Alaska Fisheries Science Center, United States National Oceanic and Atmospheric Administration (NOAA)
+ - name: NMFS Alaska Fisheries Science Center, United States National Oceanic and Atmospheric Administration, Seattle, WA, USA
    index: 3
- - name: NIASRA, School of Mathematics and Applied Statistics, University of Wollongong
+ - name: NIASRA, School of Mathematics and Applied Statistics, University of Wollongong, Wollongong, NSW, Australia
    index: 4
- - name: Rocky Mountain Research Station, United States Forest Service (USFS)
+ - name: Rocky Mountain Research Station, United States Forest Service, Boise, ID, USA
    index: 5
 citation_author: Dumelle et. al.
 date: 07 July 2024
@@ -50,7 +53,7 @@ The `SSN2` **R** package provides tools for spatial statistical modeling, parame
 
 Streams provide vital aquatic services that sustain wildlife, provide drinking and irrigation water, and support recreational and cultural activities.  Data are often collected at various locations on a stream network and used to characterize spatial patterns in stream phenomena. For example, a manager may need to know how the amount of a hazardous chemical changes throughout a stream network to inform mitigation efforts. Comprehensive formulations of spatial stream network (SSN) models are provided by @ver2010moving, @peterson2010mixed, and @ver2014ssn. The `SSN2` **R** package is designed to help users fit SSN models to their stream network data.
 
[…]
