### 2024-07-17 #6389 @mikemahoney218
Sorry, caught one more -- for Wickham 2019 can you put brackets around `{Tidyverse}`? It's capitalized in the paper title.

Otherwise, looks good to me! Note that changes in the paper don't require changes in the Zenodo archive -- JOSS makes the authoritative archive of the paper (and mints a DOI for it), while the Zenodo archive serves as the record of the software at time of acceptance. So it's fine to keep updating the paper, if needed, once the archive is made, because the "real" paper archive will be made later on 😄
