### 2024-07-18 2bcea3c
joss editorial updates
