+++ inst/joss/paper.md
@@ -102,7 +102,7 @@ ggplot() +
 
 }
 
-\caption{Middle Fork 2004 stream networks. Observed sites are represented by brown, closed circles. Prediction sites are represented by blue, closed triangles.}\label{fig:steam-network}
+\caption{Middle Fork 2004 stream networks. Observed sites are represented by brown, closed circles at various locations throughout the stream network. Prediction sites are represented by blue, closed triangles and are spaced one kilometer apart.}\label{fig:steam-network}
 \end{figure}
 
 Prior to statistical modeling, hydrologic distance matrices are created [@ver2010moving]:
@@ -292,7 +292,7 @@ ggplot() +
 
 }
 
-\caption{Predicted Middle Fork 2004 mean summer temperatures (Celsius) spaced one kilometer apart. }\label{fig:steam-preds}
+\caption{Predicted Middle Fork 2004 mean summer temperatures (Celsius) spaced one kilometer apart. As expected, temperature is predicted to be lower in areas of higher elevation.}\label{fig:steam-preds}
 \end{figure}
 
 Generalized spatial linear models for binary, count, proportion, and skewed data are available via the `ssn_glm()` function. `ssn_lm()` and `ssn_glm()` also accommodate several advanced features, which include nonspatial random effects as in `lme4` [@bates2015lme4] and `nlme` [@pinheiro2006mixed] Euclidean anisotropy [@zimmerman2024spatial], and more. In addition to modeling, simulating data on a stream network is performed via `ssn_simulate()`. 
