+++ inst/joss/paper.bib
@@ -442,7 +442,7 @@
 }
 
 @article{wickham2019welcome,
-  title={Welcome to the Tidyverse},
+  title={Welcome to the {T}idyverse},
   author={Wickham, Hadley and Averick, Mara and Bryan, Jennifer and Chang, Winston and McGowan, Lucy D'Agostino and Fran{\c{c}}ois, Romain and Grolemund, Garrett and Hayes, Alex and Henry, Lionel and Hester, Jim and others},
   journal={Journal of Open Source Software},
   volume={4},
