### 2024-07-18 #6389 @michaeldumelle
Hi @mikemahoney218, I have:

- Double checked authors and affiliations (including ORCIDs).
- Made a release of the software with the latest changes from the review. That release number is v0.2.0(JOSS). That release tag is available [linked here](https://github.com/USEPA/SSN2/releases/tag/v0.2.0(JOSS)).
- Archived the release on Zenodo at [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12770259.svg)](https://doi.org/10.5281/zenodo.12770259).
- Made sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- Made sure that the license listed for the archive is the same as the software license.

I am ready to move forward on my end; please let me know if there is anything else you need from me at this time and thank you!
