### 2024-07-18 #6389 @michaeldumelle
Hi @mikemahoney218 , in both the release tag and Zenodo archive, _both_ the source and binaries are archived. Are you suggesting that the binaries should be removed, or is it okay that they are both there? Thanks!
