### 2024-07-18 #6389 @mikemahoney218
_Oh_, sorry, Zenodo's default display only showed me the contents of the `.zip` file and not the `.tar.gz`. I think it's fine that they're both there.

@openjournals/ese-eics please ignore my concern above, this one is good to go :smile:
