### 2024-07-22 f1692d8
tidy style updates
