### 2024-07-23 #6389 @michaeldumelle
@mikemahoney218, [this paper](https://onlinelibrary.wiley.com/doi/10.1002/env.2872) came out yesterday, which details the formulation we have in the software for generalized linear models. Is it too late to add this citation (to the sentence in the paper where we currently talk about `ssn_glm()`? If it is okay to do this, I can make the change and push.
