### 2024-07-23 607c6c4
add spglm citation to joss
