+++ inst/joss/paper.bib
@@ -431,6 +431,14 @@
   doi={10.18637/jss.v056.i03}
 }
 
+@article{ver2024marginal,
+  title={Marginal inference for hierarchical generalized linear mixed models with patterned covariance matrices using the Laplace approximation},
+  author={{Ver Hoef}, Jay M and Blagg, Eryn and Dumelle, Michael and Dixon, Phillip M and Zimmerman, Dale L and Conn, Paul B},
+  journal={Environmetrics},
+  year={2024},
+  doi={10.1002/env.2872}
+}
+
 @Book{wickham2016ggplot2,
   author = {Hadley Wickham},
   title = {{ggplot2}: Elegant Graphics for Data Analysis},
