### 2024-07-23 #6389 @mikemahoney218
Looks good to me! @openjournals/ese-eics will correct me if I'm wrong, but this should be ready for processing.
