### 2024-07-25 dc0c720
updated create_netgeom function, which is now exported and makes use of overwrite argument
