### 2024-07-25 d10078d
Included call to create_netgeom in function. Removed code that converts AFV columns from numeric to text. Not required anymore
