+++ R/SSN_to_SSN2.R
@@ -4,12 +4,6 @@
 #'   created in the SSN package to an S3 \code{SSN} object used in the
 #'   SSN2 package.
 #' @param object A SpatialStreamNetwork object
-#' @param edge_additive A character vector of additive function value
-#'   column names found in edges. Default is NULL. See Details for
-#'   more information.
-#' @param site_additive A character vector of additive function value
-#'   column names found in the observed sites and prediction
-#'   sites. See Details for more information. Default is NULL.
 #' @details \command{SSN_to_SSN2()} has been made available to help users
 #'   migrate from the \code{SSN} package to the updated \code{SSN2}
 #'   package. It is used to convert existing S4 \code{SpatialStreamNetwork}
@@ -18,33 +12,12 @@
 #'   used to create an S3 \code{SSN} object from data stored locally in a .ssn
 #'   directory.
 #'
-#'   Additive function values are used to generate spatial weights for
-#'   the tail-up covariance function used in \code{ssn_glm}. The range
-#'   of additive function values are restricted to \eqn{0 \le AFV \le
-#'   1}. In the \code{SSN2} package, columns containing additive
-#'   function values are stored as text, rather than numeric
-#'   format. This prevents values less than 1 with more than 10 digits
-#'   from being truncated when writing/reading shapefiles (and their
-#'   .dbf tables). The columns containing additive function values are
-#'   specified using the \code{edge_additive} and \code{site_additive} arguments
-#'   and converted to character format in the \code{SSN} class object
-#'   returned. The arguments \code{edge_additive} and \code{site_additive}
-#'   accept a single column name in character format, or a vector
-#'   containing multiple column names. Note that, column names for
-#'   additive function values on the edges, sites, and prediction
-#'   sites may differ. If a column specified in \code{edge_additive} or
-#'   \code{site_additive} is not present, the function will return a
-#'   warning, rather than an error. Columns containing additive
-#'   function values can also be converted to text manually using the
-#'   \code{\link[base]{formatC}} function, which provides the
-#'   flexibility needed to store the values with their full precision.
-#'
-#' @return An S3 \code{SSN} class object, with additive function value
-#'   columns converted to text format.
+#' @return An S3 \code{SSN} class object.
 #'
 #' @name SSN_to_SSN2
 #' @export
-SSN_to_SSN2 <- function(object, edge_additive = NULL, site_additive = NULL) {
+#'
+SSN_to_SSN2 <- function(object) {
   .Deprecated(new = "ssn_import")
 
   if (!requireNamespace("sp", quietly = TRUE)) {
@@ -64,41 +37,43 @@ SSN_to_SSN2 <- function(object, edge_additive = NULL, site_additive = NULL) {
 
     nl.coords <- object@network.line.coords
 
-    edges$netgeom <- paste("ENETWORK",
-      paste("(",
-        paste(
-          nl.coords$NetworkID,
-          nl.coords$SegmentID,
-          nl.coords$DistanceUpstream,
-          sep = " "
-        ),
-        ")",
-        sep = ""
-      ),
-      sep = " "
-    )
+    edges <- create_netgeom(edges, type = "LINESTRING", overwrite = TRUE) 
 
-    ## Convert additive function values to text
-    if (!is.null(edge_additive)) {
[…]
