### 2024-07-25 291c737
updated functions that call, or should have called, create_netgeom() and/or were affected by update to create_netgeom
