+++ NAMESPACE
@@ -131,6 +131,7 @@ export(Torgegram)
 export(augment)
 export(copy_lsn_to_temp)
 export(covmatrix)
+export(create_netgeom)
 export(dispersion_initial)
 export(dispersion_params)
 export(euclid_initial)

[…]
