+++ man/create_netgeom.Rd
@@ -43,20 +43,13 @@ sites) geometry. For edges, the netgeom format is:
 \item{\code{'ENETWORK (netID, rid, upDist)'}}
 }
 
-The rid, upDist and netID columns must already be present in edges
-before netgeom is added. These columns are created using
-\code{link{lines_to_lsn}}, \code{\link{updist_edges}}, and
-\code{link{lsn_to_ssn}}, respectively.
-
 For observed or prediction sites, the netgeom format is:
 \itemize{
 \item{\code{'SNETWORK (netID, rid, upDist, ratio, pid, locID)'}}
 }
 
 The rid, ratio, upDist, netID, pid, and locID columns must be
-present in \code{sf_data} and are created using \code{SSNbler} package functions
-\code{sites_to_lsn}, \code{updist_sites}, and
-\code{ssn_assemble}, respectively.
+present in \code{sf_data} before netgeom is added.
 
 If \code{overwrite = TRUE} and a column named netgeom is present in
 \code{sf_data}, the data will be overwritten. Default = FALSE.
