### 2024-07-26 b21a578
update pathing for some file tests
