### 2024-07-26 #6389 @kthyng
@michaeldumelle
* Can you update the version in Zenodo to match the version here too? It currently says v1.
