### 2024-07-26 #6389 @kthyng
Congratulations on your new publication @michaeldumelle! Many thanks to @mikemahoney218 and to reviewers @fernandomayer, @k-doering-NOAA, and @fawda123 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@michaeldumelle If you'd like to review for JOSS, please register at https://reviewers.joss.theoj.org/.
