### 2024-07-26 #6389 @michaeldumelle
A final, huge thanks to @kthyng, @mikemahoney218, @fawda123, @k-doering-NOAA, and @fernandomayer for making this such an enriching process! Your hard work and feedback led me to adopt better software development practices, which notably improved both the `SSN2` software and the associated manuscript.
